Minor splitting of mutable state builder file
