+++ service/history/execution/mutable_state_builder.go
@@ -23,7 +23,6 @@ package execution
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"math/rand"
 	"time"
@@ -652,28 +651,6 @@ func (e *mutableStateBuilder) checkAndClearTimerFiredEvent(
 	return timerEvent
 }
 
-func checkAndClearTimerFiredEvent(
-	events []*types.HistoryEvent,
-	timerID string,
-) ([]*types.HistoryEvent, *types.HistoryEvent) {
-	// go over all history events. if we find a timer fired event for the given
-	// timerID, clear it
-	timerFiredIdx := -1
-	for idx, event := range events {
-		if event.GetEventType() == types.EventTypeTimerFired &&
-			event.GetTimerFiredEventAttributes().GetTimerID() == timerID {
-			timerFiredIdx = idx
-			break
-		}
-	}
-	if timerFiredIdx == -1 {
-		return events, nil
-	}
-
-	timerEvent := events[timerFiredIdx]
-	return append(events[:timerFiredIdx], events[timerFiredIdx+1:]...), timerEvent
-}
-
 func (e *mutableStateBuilder) trimEventsAfterWorkflowClose(
 	input []*types.HistoryEvent,
 ) []*types.HistoryEvent {
@@ -961,135 +938,6 @@ func (e *mutableStateBuilder) SetQueryRegistry(queryRegistry query.Registry) {
 	e.queryRegistry = queryRegistry
 }
 
-func (e *mutableStateBuilder) GetActivityScheduledEvent(
-	ctx context.Context,
-	scheduleEventID int64,
-) (*types.HistoryEvent, error) {
-
-	ai, ok := e.pendingActivityInfoIDs[scheduleEventID]
-	if !ok {
-		return nil, ErrMissingActivityInfo
-	}
-
-	// Needed for backward compatibility reason
-	if ai.ScheduledEvent != nil {
-		return ai.ScheduledEvent, nil
-	}
-
-	currentBranchToken, err := e.GetCurrentBranchToken()
-	if err != nil {
-		return nil, err
-	}
-	scheduledEvent, err := e.eventsCache.GetEvent(
-		ctx,
-		e.shard.GetShardID(),
-		e.executionInfo.DomainID,
-		e.executionInfo.WorkflowID,
-		e.executionInfo.RunID,
-		ai.ScheduledEventBatchID,
-		ai.ScheduleID,
-		currentBranchToken,
-	)
-	if err != nil {
-		// do not return the original error
-		// since original error can be of type entity not exists
-		// which can cause task processing side to fail silently
-		// However, if the error is a persistence transient error,
-		// we return the original error, because we fail to get
-		// the event because of failure from database
-		if persistence.IsTransientError(err) {
-			return nil, err
-		}
-		return nil, ErrMissingActivityScheduledEvent
-	}
-	return scheduledEvent, nil
-}
-
-// GetActivityInfo gives details about an activity that is currently in progress.
-func (e *mutableStateBuilder) GetActivityInfo(
-	scheduleEventID int64,
-) (*persistence.ActivityInfo, bool) {
-
-	ai, ok := e.pendingActivityInfoIDs[scheduleEventID]
-	return ai, ok
-}
-
-// GetActivityByActivityID gives details about an activity that is currently in progress.
-func (e *mutableStateBuilder) GetActivityByActivityID(
-	activityID string,
-) (*persistence.ActivityInfo, bool) {
-
-	eventID, ok := e.pendingActivityIDToEventID[activityID]
-	if !ok {
-		return nil, false
-	}
-	return e.GetActivityInfo(eventID)
-}
-
-// GetChildExecutionInfo gives details about a child execution that is currently in progress.
-func (e *mutableStateBuilder) GetChildExecutionInfo(
-	initiatedEventID int64,
-) (*persistence.ChildExecutionInfo, bool) {
-
-	ci, ok := e.pendingChildExecutionInfoIDs[initiatedEventID]
-	return ci, ok
-}
-
-// GetChildExecutionInitiatedEvent reads out the ChildExecutionInitiatedEvent from mutable state for in-progress child
-// executions
-func (e *mutableStateBuilder) GetChildExecutionInitiatedEvent(
-	ctx context.Context,
-	initiatedEventID int64,
-) (*types.HistoryEvent, error) {
-
-	ci, ok := e.pendingChildExecutionInfoIDs[initiatedEventID]
-	if !ok {
-		return nil, ErrMissingChildWorkflowInfo
-	}
-
-	// Needed for backward compatibility reason
-	if ci.InitiatedEvent != nil {
-		return ci.InitiatedEvent, nil
-	}
-
-	currentBranchToken, err := e.GetCurrentBranchToken()
-	if err != nil {
-		return nil, err
-	}
-	initiatedEvent, err := e.eventsCache.GetEvent(
-		ctx,
-		e.shard.GetShardID(),
-		e.executionInfo.DomainID,
-		e.executionInfo.WorkflowID,
-		e.executionInfo.RunID,
-		ci.InitiatedEventBatchID,
-		ci.InitiatedID,
-		currentBranchToken,
-	)
-	if err != nil {
-		// do not return the original error
-		// since original error can be of type entity not exists
-		// which can cause task processing side to fail silently
-		// However, if the error is a persistence transient error,
-		// we return the original error, because we fail to get
-		// the event because of failure from database
-		if persistence.IsTransientError(err) {
-			return nil, err
-		}
-		return nil, ErrMissingChildWorkflowInitiatedEvent
-	}
-	return initiatedEvent, nil
-}
-
-// GetRequestCancelInfo gives details about a request cancellation that is currently in progress.
-func (e *mutableStateBuilder) GetRequestCancelInfo(
-	initiatedEventID int64,
-) (*persistence.RequestCancelInfo, bool) {
-
-	ri, ok := e.pendingRequestCancelInfoIDs[initiatedEventID]
-	return ri, ok
-}
-
 func (e *mutableStateBuilder) GetRetryBackoffDuration(
 	errReason string,
 ) time.Duration {
@@ -1138,67 +986,6 @@ func (e *mutableStateBuilder) GetCronBackoffDuration(
 	return backoff.GetBackoffForNextSchedule(sched, executionTime, e.timeSource.Now(), jitterStartSeconds)
 }
 
-// GetSignalInfo get details about a signal request that is currently in progress.
-func (e *mutableStateBuilder) GetSignalInfo(
-	initiatedEventID int64,
-) (*persistence.SignalInfo, bool) {
-
-	ri, ok := e.pendingSignalInfoIDs[initiatedEventID]
-	return ri, ok
-}
-
-// GetCompletionEvent retrieves the workflow completion event from mutable state
-func (e *mutableStateBuilder) GetCompletionEvent(
-	ctx context.Context,
-) (*types.HistoryEvent, error) {
-	if e.executionInfo.State != persistence.WorkflowStateCompleted {
-		return nil, ErrMissingWorkflowCompletionEvent
-	}
-
-	// Needed for backward compatibility reason
-	if e.executionInfo.CompletionEvent != nil {
-		return e.executionInfo.CompletionEvent, nil
-	}
-
-	// Needed for backward compatibility reason
-	if e.executionInfo.CompletionEventBatchID == common.EmptyEventID {
-		return nil, ErrMissingWorkflowCompletionEvent
-	}
-
-	currentBranchToken, err := e.GetCurrentBranchToken()
-	if err != nil {
-		return nil, err
-	}
-
-	// Completion EventID is always one less than NextEventID after workflow is completed
-	completionEventID := e.executionInfo.NextEventID - 1
-	firstEventID := e.executionInfo.CompletionEventBatchID
-	completionEvent, err := e.eventsCache.GetEvent(
-		ctx,
-		e.shard.GetShardID(),
-		e.executionInfo.DomainID,
-		e.executionInfo.WorkflowID,
-		e.executionInfo.RunID,
-		firstEventID,
-		completionEventID,
-		currentBranchToken,
-	)
-	if err != nil {
-		// do not return the original error
-		// since original error can be of type entity not exists
-		// which can cause task processing side to fail silently
-		// However, if the error is a persistence transient error,
-		// we return the original error, because we fail to get
-		// the event because of failure from database
-		if persistence.IsTransientError(err) {
-			return nil, err
-		}
-		return nil, ErrMissingWorkflowCompletionEvent
-	}
-
-	return completionEvent, nil
-}
-
 // GetStartEvent retrieves the workflow start event from mutable state
 func (e *mutableStateBuilder) GetStartEvent(
 	ctx context.Context,
@@ -1234,27 +1021,6 @@ func (e *mutableStateBuilder) GetStartEvent(
 	return startEvent, nil
 }
 
-// DeletePendingChildExecution deletes details about a ChildExecutionInfo.
-func (e *mutableStateBuilder) DeletePendingChildExecution(
-	initiatedEventID int64,
-) error {
-
-	if _, ok := e.pendingChildExecutionInfoIDs[initiatedEventID]; ok {
-		delete(e.pendingChildExecutionInfoIDs, initiatedEventID)
-	} else {
-		e.logError(
-			fmt.Sprintf("unable to find child workflow event ID: %v in mutable state", initiatedEventID),
-			tag.ErrorTypeInvalidMutableStateAction,
-		)
-		// log data inconsistency instead of returning an error
-		e.logDataInconsistency()
-	}
-
-	delete(e.updateChildExecutionInfos, initiatedEventID)
-	e.deleteChildExecutionInfos[initiatedEventID] = struct{}{}
-	return nil
-}
-
 // DeletePendingRequestCancel deletes details about a RequestCancelInfo.
 func (e *mutableStateBuilder) DeletePendingRequestCancel(
 	initiatedEventID int64,
@@ -1273,27 +1039,6 @@ func (e *mutableStateBuilder) DeletePendingRequestCancel(
 
 	delete(e.updateRequestCancelInfos, initiatedEventID)
 	e.deleteRequestCancelInfos[initiatedEventID] = struct{}{}
-	return nil
-}
-
-// DeletePendingSignal deletes details about a SignalInfo
-func (e *mutableStateBuilder) DeletePendingSignal(
-	initiatedEventID int64,
-) error {
-
-	if _, ok := e.pendingSignalInfoIDs[initiatedEventID]; ok {
-		delete(e.pendingSignalInfoIDs, initiatedEventID)
-	} else {
-		e.logError(
-			fmt.Sprintf("unable to find signal external workflow event ID: %v in mutable state", initiatedEventID),
-			tag.ErrorTypeInvalidMutableStateAction,
-		)
-		// log data inconsistency instead of returning an error
-		e.logDataInconsistency()
-	}
-
-	delete(e.updateSignalInfos, initiatedEventID)
-	e.deleteSignalInfos[initiatedEventID] = struct{}{}
 	return nil
 }
 
@@ -1318,113 +1063,6 @@ func (e *mutableStateBuilder) HasParentExecution() bool {
 	return e.executionInfo.ParentDomainID != "" && e.executionInfo.ParentWorkflowID != ""
 }
 
-func (e *mutableStateBuilder) UpdateActivityProgress(
-	ai *persistence.ActivityInfo,
-	request *types.RecordActivityTaskHeartbeatRequest,
-) {
-	ai.Version = e.GetCurrentVersion()
-	ai.Details = request.Details
-	ai.LastHeartBeatUpdatedTime = e.timeSource.Now()
-	e.updateActivityInfos[ai.ScheduleID] = ai
-	e.syncActivityTasks[ai.ScheduleID] = struct{}{}
-}
-
-// ReplicateActivityInfo replicate the necessary activity information
-func (e *mutableStateBuilder) ReplicateActivityInfo(
-	request *types.SyncActivityRequest,
-	resetActivityTimerTaskStatus bool,
-) error {
-	ai, ok := e.pendingActivityInfoIDs[request.GetScheduledID()]
-	if !ok {
-		e.logError(
-			fmt.Sprintf("unable to find activity event ID: %v in mutable state", request.GetScheduledID()),
-			tag.ErrorTypeInvalidMutableStateAction,
-		)
-		return ErrMissingActivityInfo
-	}
-
-	ai.Version = request.GetVersion()
-	ai.ScheduledTime = time.Unix(0, request.GetScheduledTime())
-	ai.StartedID = request.GetStartedID()
-	ai.LastHeartBeatUpdatedTime = time.Unix(0, request.GetLastHeartbeatTime())
-	if ai.StartedID == common.EmptyEventID {
-		ai.StartedTime = time.Time{}
-	} else {
-		ai.StartedTime = time.Unix(0, request.GetStartedTime())
-	}
-	ai.Details = request.GetDetails()
-	ai.Attempt = request.GetAttempt()
-	ai.LastFailureReason = request.GetLastFailureReason()
-	ai.LastWorkerIdentity = request.GetLastWorkerIdentity()
-	ai.LastFailureDetails = request.GetLastFailureDetails()
-
-	if resetActivityTimerTaskStatus {
-		ai.TimerTaskStatus = TimerTaskStatusNone
-	}
-
-	e.updateActivityInfos[ai.ScheduleID] = ai
-	return nil
-}
-
-// UpdateActivity updates an activity
-func (e *mutableStateBuilder) UpdateActivity(
-	ai *persistence.ActivityInfo,
-) error {
-
-	if _, ok := e.pendingActivityInfoIDs[ai.ScheduleID]; !ok {
-		e.logError(
-			fmt.Sprintf("unable to find activity ID: %v in mutable state", ai.ActivityID),
-			tag.ErrorTypeInvalidMutableStateAction,
-		)
-		return ErrMissingActivityInfo
-	}
-
-	e.pendingActivityInfoIDs[ai.ScheduleID] = ai
-	e.updateActivityInfos[ai.ScheduleID] = ai
-	return nil
-}
-
-// DeleteActivity deletes details about an activity.
-func (e *mutableStateBuilder) DeleteActivity(
-	scheduleEventID int64,
-) error {
-
-	if activityInfo, ok := e.pendingActivityInfoIDs[scheduleEventID]; ok {
-		delete(e.pendingActivityInfoIDs, scheduleEventID)
-
-		if _, ok = e.pendingActivityIDToEventID[activityInfo.ActivityID]; ok {
-			delete(e.pendingActivityIDToEventID, activityInfo.ActivityID)
-		} else {
-			e.logError(
-				fmt.Sprintf("unable to find activity ID: %v in mutable state", activityInfo.ActivityID),
-				tag.ErrorTypeInvalidMutableStateAction,
-			)
-			// log data inconsistency instead of returning an error
-			e.logDataInconsistency()
-		}
-	} else {
-		e.logError(
-			fmt.Sprintf("unable to find activity event id: %v in mutable state", scheduleEventID),
-			tag.ErrorTypeInvalidMutableStateAction,
-		)
-		// log data inconsistency instead of returning an error
-		e.logDataInconsistency()
-	}
-
-	delete(e.updateActivityInfos, scheduleEventID)
-	e.deleteActivityInfos[scheduleEventID] = struct{}{}
-	return nil
-}
-
-// GetUserTimerInfo gives details about a user timer.
-func (e *mutableStateBuilder) GetUserTimerInfo(
-	timerID string,
-) (*persistence.TimerInfo, bool) {
-
-	timerInfo, ok := e.pendingTimerInfoIDs[timerID]
-	return timerInfo, ok
-}
-
 // GetUserTimerInfoByEventID gives details about a user timer.
 func (e *mutableStateBuilder) GetUserTimerInfoByEventID(
 	startEventID int64,
@@ -1437,114 +1075,8 @@ func (e *mutableStateBuilder) GetUserTimerInfoByEventID(
 	return e.GetUserTimerInfo(timerID)
 }
 
-// UpdateUserTimer updates the user timer in progress.
-func (e *mutableStateBuilder) UpdateUserTimer(
-	ti *persistence.TimerInfo,
-) error {
-
-	timerID, ok := e.pendingTimerEventIDToID[ti.StartedID]
-	if !ok {
-		e.logError(
-			fmt.Sprintf("unable to find timer event ID: %v in mutable state", ti.StartedID),
-			tag.ErrorTypeInvalidMutableStateAction,
-		)
-		return ErrMissingTimerInfo
-	}
-
-	if _, ok := e.pendingTimerInfoIDs[timerID]; !ok {
-		e.logError(
-			fmt.Sprintf("unable to find timer ID: %v in mutable state", timerID),
-			tag.ErrorTypeInvalidMutableStateAction,
-		)
-		return ErrMissingTimerInfo
-	}
-
-	e.pendingTimerInfoIDs[ti.TimerID] = ti
-	e.updateTimerInfos[ti.TimerID] = ti
-	return nil
-}
-
-// DeleteUserTimer deletes an user timer.
-func (e *mutableStateBuilder) DeleteUserTimer(
-	timerID string,
-) error {
-
-	if timerInfo, ok := e.pendingTimerInfoIDs[timerID]; ok {
-		delete(e.pendingTimerInfoIDs, timerID)
-
-		if _, ok = e.pendingTimerEventIDToID[timerInfo.StartedID]; ok {
-			delete(e.pendingTimerEventIDToID, timerInfo.StartedID)
-		} else {
-			e.logError(
-				fmt.Sprintf("unable to find timer event ID: %v in mutable state", timerID),
-				tag.ErrorTypeInvalidMutableStateAction,
-			)
-			// log data inconsistency instead of returning an error
-			e.logDataInconsistency()
-		}
-	} else {
-		e.logError(
-			fmt.Sprintf("unable to find timer ID: %v in mutable state", timerID),
-			tag.ErrorTypeInvalidMutableStateAction,
-		)
-		// log data inconsistency instead of returning an error
-		e.logDataInconsistency()
-	}
-
-	delete(e.updateTimerInfos, timerID)
-	e.deleteTimerInfos[timerID] = struct{}{}
-	return nil
-}
-
-// GetDecisionInfo returns details about the in-progress decision task
-func (e *mutableStateBuilder) GetDecisionInfo(
-	scheduleEventID int64,
-) (*DecisionInfo, bool) {
-	return e.decisionTaskManager.GetDecisionInfo(scheduleEventID)
-}
-
-func (e *mutableStateBuilder) GetDecisionScheduleToStartTimeout() time.Duration {
-	return e.decisionTaskManager.GetDecisionScheduleToStartTimeout()
-}
-
-func (e *mutableStateBuilder) GetPendingActivityInfos() map[int64]*persistence.ActivityInfo {
-	return e.pendingActivityInfoIDs
-}
-
-func (e *mutableStateBuilder) GetPendingTimerInfos() map[string]*persistence.TimerInfo {
-	return e.pendingTimerInfoIDs
-}
-
-func (e *mutableStateBuilder) GetPendingChildExecutionInfos() map[int64]*persistence.ChildExecutionInfo {
-	return e.pendingChildExecutionInfoIDs
-}
-
 func (e *mutableStateBuilder) GetPendingRequestCancelExternalInfos() map[int64]*persistence.RequestCancelInfo {
 	return e.pendingRequestCancelInfoIDs
-}
-
-func (e *mutableStateBuilder) GetPendingSignalExternalInfos() map[int64]*persistence.SignalInfo {
-	return e.pendingSignalInfoIDs
-}
-
-func (e *mutableStateBuilder) HasProcessedOrPendingDecision() bool {
-	return e.decisionTaskManager.HasProcessedOrPendingDecision()
-}
-
-func (e *mutableStateBuilder) HasPendingDecision() bool {
-	return e.decisionTaskManager.HasPendingDecision()
-}
-
-func (e *mutableStateBuilder) GetPendingDecision() (*DecisionInfo, bool) {
-	return e.decisionTaskManager.GetPendingDecision()
-}
-
-func (e *mutableStateBuilder) HasInFlightDecision() bool {
-	return e.decisionTaskManager.HasInFlightDecision()
-}
-
-func (e *mutableStateBuilder) GetInFlightDecision() (*DecisionInfo, bool) {
-	return e.decisionTaskManager.GetInFlightDecision()
 }
 
 func (e *mutableStateBuilder) HasBufferedEvents() bool {
@@ -1559,24 +1091,6 @@ func (e *mutableStateBuilder) HasBufferedEvents() bool {
 	}
 
 	return false
-}
-
-// UpdateDecision updates a decision task.
-func (e *mutableStateBuilder) UpdateDecision(
-	decision *DecisionInfo,
-) {
-	e.decisionTaskManager.UpdateDecision(decision)
-}
-
-// DeleteDecision deletes a decision task.
-func (e *mutableStateBuilder) DeleteDecision() {
-	e.decisionTaskManager.DeleteDecision()
-}
-
-func (e *mutableStateBuilder) FailDecision(
-	incrementAttempt bool,
-) {
-	e.decisionTaskManager.FailDecision(incrementAttempt)
 }
 
 func (e *mutableStateBuilder) ClearStickyness() {
@@ -1607,1426 +1121,6 @@ func (e *mutableStateBuilder) IsWorkflowExecutionRunning() bool {
 	return e.executionInfo.IsRunning()
 }
 
-func (e *mutableStateBuilder) IsWorkflowCompleted() bool {
-	return e.executionInfo.State == persistence.WorkflowStateCompleted
-}
-
-func (e *mutableStateBuilder) IsCancelRequested() (bool, string) {
-	if e.executionInfo.CancelRequested {
-		return e.executionInfo.CancelRequested, e.executionInfo.CancelRequestID
-	}
-
-	return false, ""
-}
-
-func (e *mutableStateBuilder) IsSignalRequested(
-	requestID string,
-) bool {
-
-	if _, ok := e.pendingSignalRequestedIDs[requestID]; ok {
-		return true
-	}
-	return false
-}
-
-func (e *mutableStateBuilder) AddSignalRequested(
-	requestID string,
-) {
-
-	if e.pendingSignalRequestedIDs == nil {
-		e.pendingSignalRequestedIDs = make(map[string]struct{})
-	}
-	if e.updateSignalRequestedIDs == nil {
-		e.updateSignalRequestedIDs = make(map[string]struct{})
-	}
-	e.pendingSignalRequestedIDs[requestID] = struct{}{} // add requestID to set
-	e.updateSignalRequestedIDs[requestID] = struct{}{}
-}
-
-func (e *mutableStateBuilder) DeleteSignalRequested(
-	requestID string,
-) {
-
-	delete(e.pendingSignalRequestedIDs, requestID)
-	delete(e.updateSignalRequestedIDs, requestID)
-	e.deleteSignalRequestedIDs[requestID] = struct{}{}
-}
-
-func (e *mutableStateBuilder) addWorkflowExecutionStartedEventForContinueAsNew(
-	parentExecutionInfo *types.ParentExecutionInfo,
-	execution types.WorkflowExecution,
-	previousExecutionState MutableState,
-	attributes *types.ContinueAsNewWorkflowExecutionDecisionAttributes,
-	firstRunID string,
-	firstScheduledTime time.Time,
-) (*types.HistoryEvent, error) {
-
-	previousExecutionInfo := previousExecutionState.GetExecutionInfo()
-	taskList := previousExecutionInfo.TaskList
-	if attributes.TaskList != nil {
-		taskList = attributes.TaskList.GetName()
-	}
-	tl := &types.TaskList{}
-	tl.Name = taskList
-
-	workflowType := previousExecutionInfo.WorkflowTypeName
-	if attributes.WorkflowType != nil {
-		workflowType = attributes.WorkflowType.GetName()
-	}
-	wType := &types.WorkflowType{}
-	wType.Name = workflowType
-
-	decisionTimeout := previousExecutionInfo.DecisionStartToCloseTimeout
-	if attributes.TaskStartToCloseTimeoutSeconds != nil {
-		decisionTimeout = attributes.GetTaskStartToCloseTimeoutSeconds()
-	}
-
-	createRequest := &types.StartWorkflowExecutionRequest{
-		RequestID:                           uuid.New(),
-		Domain:                              e.domainEntry.GetInfo().Name,
-		WorkflowID:                          execution.WorkflowID,
-		TaskList:                            tl,
-		WorkflowType:                        wType,
-		TaskStartToCloseTimeoutSeconds:      common.Int32Ptr(decisionTimeout),
-		ExecutionStartToCloseTimeoutSeconds: attributes.ExecutionStartToCloseTimeoutSeconds,
-		Input:                               attributes.Input,
-		Header:                              attributes.Header,
-		RetryPolicy:                         attributes.RetryPolicy,
-		CronSchedule:                        attributes.CronSchedule,
-		Memo:                                attributes.Memo,
-		SearchAttributes:                    attributes.SearchAttributes,
-		JitterStartSeconds:                  attributes.JitterStartSeconds,
-	}
-
-	req := &types.HistoryStartWorkflowExecutionRequest{
-		DomainUUID:                      e.domainEntry.GetInfo().ID,
-		StartRequest:                    createRequest,
-		ParentExecutionInfo:             parentExecutionInfo,
-		LastCompletionResult:            attributes.LastCompletionResult,
-		ContinuedFailureReason:          attributes.FailureReason,
-		ContinuedFailureDetails:         attributes.FailureDetails,
-		ContinueAsNewInitiator:          attributes.Initiator,
-		FirstDecisionTaskBackoffSeconds: attributes.BackoffStartIntervalInSeconds,
-		PartitionConfig:                 previousExecutionInfo.PartitionConfig,
-	}
-
-	// if ContinueAsNew as Cron or decider, recalculate the expiration timestamp and set attempts to 0
-	req.Attempt = 0
-	if attributes.RetryPolicy != nil && attributes.RetryPolicy.GetExpirationIntervalInSeconds() > 0 {
-		// expirationTime calculates from first decision task schedule to the end of the workflow
-		expirationInSeconds := attributes.RetryPolicy.GetExpirationIntervalInSeconds() + req.GetFirstDecisionTaskBackoffSeconds()
-		expirationTime := e.timeSource.Now().Add(time.Second * time.Duration(expirationInSeconds))
-		req.ExpirationTimestamp = common.Int64Ptr(expirationTime.UnixNano())
-	}
-	// if ContinueAsNew as retry use the same expiration timestamp and increment attempts from previous execution state
-	if attributes.GetInitiator() == types.ContinueAsNewInitiatorRetryPolicy {
-		req.Attempt = previousExecutionState.GetExecutionInfo().Attempt + 1
-		expirationTime := previousExecutionState.GetExecutionInfo().ExpirationTime
-		if !expirationTime.IsZero() {
-			req.ExpirationTimestamp = common.Int64Ptr(expirationTime.UnixNano())
-		}
-	}
-
-	// History event only has domainName so domainID has to be passed in explicitly to update the mutable state
-	var parentDomainID *string
-	if parentExecutionInfo != nil {
-		parentDomainID = &parentExecutionInfo.DomainUUID
-	}
-
-	event := e.hBuilder.AddWorkflowExecutionStartedEvent(req, previousExecutionInfo, firstRunID, execution.GetRunID(), firstScheduledTime)
-	if err := e.ReplicateWorkflowExecutionStartedEvent(
-		parentDomainID,
-		execution,
-		createRequest.GetRequestID(),
-		event,
-		false,
-	); err != nil {
-		return nil, err
-	}
-
-	if err := e.SetHistoryTree(e.GetExecutionInfo().RunID); err != nil {
-		return nil, err
-	}
-
-	if err := e.AddFirstDecisionTaskScheduled(
-		event,
-	); err != nil {
-		return nil, err
-	}
-
-	return event, nil
-}
-
-func (e *mutableStateBuilder) AddWorkflowExecutionStartedEvent(
-	execution types.WorkflowExecution,
-	startRequest *types.HistoryStartWorkflowExecutionRequest,
-) (*types.HistoryEvent, error) {
-
-	opTag := tag.WorkflowActionWorkflowStarted
-	if err := e.checkMutability(opTag); err != nil {
-		return nil, err
-	}
-
-	request := startRequest.StartRequest
-	eventID := e.GetNextEventID()
-	if eventID != common.FirstEventID {
-		e.logger.Warn(mutableStateInvalidHistoryActionMsg, opTag,
-			tag.WorkflowEventID(eventID),
-			tag.ErrorTypeInvalidHistoryAction)
-		return nil, e.createInternalServerError(opTag)
-	}
-
-	event := e.hBuilder.AddWorkflowExecutionStartedEvent(startRequest, nil, execution.GetRunID(), execution.GetRunID(),
-		time.Now())
-
-	var parentDomainID *string
-	if startRequest.ParentExecutionInfo != nil {
-		parentDomainID = &startRequest.ParentExecutionInfo.DomainUUID
-	}
-	if err := e.ReplicateWorkflowExecutionStartedEvent(
-		parentDomainID,
-		execution,
-		request.GetRequestID(),
-		event,
-		false); err != nil {
-		return nil, err
-	}
-
-	return event, nil
-}
-
-func (e *mutableStateBuilder) ReplicateWorkflowExecutionStartedEvent(
-	parentDomainID *string,
-	execution types.WorkflowExecution,
-	requestID string,
-	startEvent *types.HistoryEvent,
-	generateDelayedDecisionTasks bool,
-) error {
-
-	event := startEvent.WorkflowExecutionStartedEventAttributes
-	if event.GetRequestID() != "" {
-		// prefer requestID from history event, ideally we should remove the requestID parameter
-		// removing it may or may not be backward compatible, so keep it now
-		requestID = event.GetRequestID()
-	}
-	e.executionInfo.CreateRequestID = requestID
-	e.insertWorkflowRequest(persistence.WorkflowRequest{
-		RequestID:   requestID,
-		Version:     startEvent.Version,
-		RequestType: persistence.WorkflowRequestTypeStart,
-	})
-	e.executionInfo.DomainID = e.domainEntry.GetInfo().ID
-	e.executionInfo.WorkflowID = execution.GetWorkflowID()
-	e.executionInfo.RunID = execution.GetRunID()
-	e.executionInfo.FirstExecutionRunID = event.GetFirstExecutionRunID()
-	if e.executionInfo.FirstExecutionRunID == "" {
-		e.executionInfo.FirstExecutionRunID = execution.GetRunID()
-	}
-	e.executionInfo.TaskList = event.TaskList.GetName()
-	e.executionInfo.WorkflowTypeName = event.WorkflowType.GetName()
-	e.executionInfo.WorkflowTimeout = event.GetExecutionStartToCloseTimeoutSeconds()
-	e.executionInfo.DecisionStartToCloseTimeout = event.GetTaskStartToCloseTimeoutSeconds()
-	e.executionInfo.StartTimestamp = e.unixNanoToTime(startEvent.GetTimestamp())
-
-	if err := e.UpdateWorkflowStateCloseStatus(
-		persistence.WorkflowStateCreated,
-		persistence.WorkflowCloseStatusNone,
-	); err != nil {
-		return err
-	}
-	e.executionInfo.LastProcessedEvent = common.EmptyEventID
-	e.executionInfo.LastFirstEventID = startEvent.ID
-
-	e.executionInfo.DecisionVersion = common.EmptyVersion
-	e.executionInfo.DecisionScheduleID = common.EmptyEventID
-	e.executionInfo.DecisionStartedID = common.EmptyEventID
-	e.executionInfo.DecisionRequestID = common.EmptyUUID
-	e.executionInfo.DecisionTimeout = 0
-
-	e.executionInfo.CronSchedule = event.GetCronSchedule()
-
-	if parentDomainID != nil {
-		e.executionInfo.ParentDomainID = *parentDomainID
-	}
-	if event.ParentWorkflowExecution != nil {
-		e.executionInfo.ParentWorkflowID = event.ParentWorkflowExecution.GetWorkflowID()
-		e.executionInfo.ParentRunID = event.ParentWorkflowExecution.GetRunID()
-	}
-	if event.ParentInitiatedEventID != nil {
-		e.executionInfo.InitiatedID = event.GetParentInitiatedEventID()
-	} else {
-		e.executionInfo.InitiatedID = common.EmptyEventID
-	}
-
-	e.executionInfo.Attempt = event.GetAttempt()
-	if event.GetExpirationTimestamp() != 0 {
-		e.executionInfo.ExpirationTime = time.Unix(0, event.GetExpirationTimestamp())
-	}
-	if event.RetryPolicy != nil {
-		e.executionInfo.HasRetryPolicy = true
-		e.executionInfo.BackoffCoefficient = event.RetryPolicy.GetBackoffCoefficient()
-		e.executionInfo.ExpirationSeconds = event.RetryPolicy.GetExpirationIntervalInSeconds()
-		e.executionInfo.InitialInterval = event.RetryPolicy.GetInitialIntervalInSeconds()
-		e.executionInfo.MaximumAttempts = event.RetryPolicy.GetMaximumAttempts()
-		e.executionInfo.MaximumInterval = event.RetryPolicy.GetMaximumIntervalInSeconds()
-		e.executionInfo.NonRetriableErrors = event.RetryPolicy.NonRetriableErrorReasons
-	}
-
-	e.executionInfo.AutoResetPoints = rolloverAutoResetPointsWithExpiringTime(
-		event.GetPrevAutoResetPoints(),
-		event.GetContinuedExecutionRunID(),
-		startEvent.GetTimestamp(),
-		e.domainEntry.GetRetentionDays(e.executionInfo.WorkflowID),
-	)
-
-	if event.Memo != nil {
-		e.executionInfo.Memo = event.Memo.GetFields()
-	}
-	if event.SearchAttributes != nil {
-		e.executionInfo.SearchAttributes = event.SearchAttributes.GetIndexedFields()
-	}
-	e.executionInfo.PartitionConfig = event.PartitionConfig
-
-	e.writeEventToCache(startEvent)
-
-	if err := e.taskGenerator.GenerateWorkflowStartTasks(e.unixNanoToTime(startEvent.GetTimestamp()), startEvent); err != nil {
-		return err
-	}
-	if err := e.taskGenerator.GenerateRecordWorkflowStartedTasks(startEvent); err != nil {
-		return err
-	}
-	if generateDelayedDecisionTasks && event.GetFirstDecisionTaskBackoffSeconds() > 0 {
-		if err := e.taskGenerator.GenerateDelayedDecisionTasks(startEvent); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func (e *mutableStateBuilder) AddFirstDecisionTaskScheduled(
-	startEvent *types.HistoryEvent,
-) error {
-	opTag := tag.WorkflowActionDecisionTaskScheduled
-	if err := e.checkMutability(opTag); err != nil {
-		return err
-	}
-	return e.decisionTaskManager.AddFirstDecisionTaskScheduled(startEvent)
-}
-
-func (e *mutableStateBuilder) AddDecisionTaskScheduledEvent(
-	bypassTaskGeneration bool,
-) (*DecisionInfo, error) {
-	opTag := tag.WorkflowActionDecisionTaskScheduled
-	if err := e.checkMutability(opTag); err != nil {
-		return nil, err
-	}
-	return e.decisionTaskManager.AddDecisionTaskScheduledEvent(bypassTaskGeneration)
-}
-
-// originalScheduledTimestamp is to record the first scheduled decision during decision heartbeat.
-func (e *mutableStateBuilder) AddDecisionTaskScheduledEventAsHeartbeat(
-	bypassTaskGeneration bool,
-	originalScheduledTimestamp int64,
-) (*DecisionInfo, error) {
-	opTag := tag.WorkflowActionDecisionTaskScheduled
-	if err := e.checkMutability(opTag); err != nil {
-		return nil, err
-	}
-	return e.decisionTaskManager.AddDecisionTaskScheduledEventAsHeartbeat(bypassTaskGeneration, originalScheduledTimestamp)
-}
-
-func (e *mutableStateBuilder) ReplicateTransientDecisionTaskScheduled() error {
-	return e.decisionTaskManager.ReplicateTransientDecisionTaskScheduled()
-}
-
-func (e *mutableStateBuilder) ReplicateDecisionTaskScheduledEvent(
-	version int64,
-	scheduleID int64,
-	taskList string,
-	startToCloseTimeoutSeconds int32,
-	attempt int64,
-	scheduleTimestamp int64,
-	originalScheduledTimestamp int64,
-	bypassTaskGeneration bool,
-) (*DecisionInfo, error) {
-	return e.decisionTaskManager.ReplicateDecisionTaskScheduledEvent(version, scheduleID, taskList, startToCloseTimeoutSeconds, attempt, scheduleTimestamp, originalScheduledTimestamp, bypassTaskGeneration)
-}
-
-func (e *mutableStateBuilder) AddDecisionTaskStartedEvent(
-	scheduleEventID int64,
-	requestID string,
-	request *types.PollForDecisionTaskRequest,
-) (*types.HistoryEvent, *DecisionInfo, error) {
-	opTag := tag.WorkflowActionDecisionTaskStarted
-	if err := e.checkMutability(opTag); err != nil {
-		return nil, nil, err
-	}
-	return e.decisionTaskManager.AddDecisionTaskStartedEvent(scheduleEventID, requestID, request)
-}
-
-func (e *mutableStateBuilder) ReplicateDecisionTaskStartedEvent(
-	decision *DecisionInfo,
-	version int64,
-	scheduleID int64,
-	startedID int64,
-	requestID string,
-	timestamp int64,
-) (*DecisionInfo, error) {
-
-	return e.decisionTaskManager.ReplicateDecisionTaskStartedEvent(decision, version, scheduleID, startedID, requestID, timestamp)
-}
-
-func (e *mutableStateBuilder) CreateTransientDecisionEvents(
-	decision *DecisionInfo,
-	identity string,
-) (*types.HistoryEvent, *types.HistoryEvent) {
-	return e.decisionTaskManager.CreateTransientDecisionEvents(decision, identity)
-}
-
-// add BinaryCheckSum for the first decisionTaskCompletedID for auto-reset
-func (e *mutableStateBuilder) addBinaryCheckSumIfNotExists(
-	event *types.HistoryEvent,
-	maxResetPoints int,
-) error {
-	binChecksum := event.GetDecisionTaskCompletedEventAttributes().GetBinaryChecksum()
-	if len(binChecksum) == 0 {
-		return nil
-	}
-	exeInfo := e.executionInfo
-	var currResetPoints []*types.ResetPointInfo
-	if exeInfo.AutoResetPoints != nil && exeInfo.AutoResetPoints.Points != nil {
-		currResetPoints = e.executionInfo.AutoResetPoints.Points
-	} else {
-		currResetPoints = make([]*types.ResetPointInfo, 0, 1)
-	}
-
-	// List of all recent binary checksums associated with the types.
-	var recentBinaryChecksums []string
-
-	for _, rp := range currResetPoints {
-		recentBinaryChecksums = append(recentBinaryChecksums, rp.GetBinaryChecksum())
-		if rp.GetBinaryChecksum() == binChecksum {
-			// this checksum already exists
-			return nil
-		}
-	}
-
-	recentBinaryChecksums, currResetPoints = trimBinaryChecksums(recentBinaryChecksums, currResetPoints, maxResetPoints)
-
-	// Adding current version of the binary checksum.
-	recentBinaryChecksums = append(recentBinaryChecksums, binChecksum)
-
-	resettable := true
-	err := e.CheckResettable()
-	if err != nil {
-		resettable = false
-	}
-	info := &types.ResetPointInfo{
-		BinaryChecksum:           binChecksum,
-		RunID:                    exeInfo.RunID,
-		FirstDecisionCompletedID: event.ID,
-		CreatedTimeNano:          common.Int64Ptr(e.timeSource.Now().UnixNano()),
-		Resettable:               resettable,
-	}
-	currResetPoints = append(currResetPoints, info)
-	exeInfo.AutoResetPoints = &types.ResetPoints{
-		Points: currResetPoints,
-	}
-	bytes, err := json.Marshal(recentBinaryChecksums)
-	if err != nil {
-		return err
-	}
-	if exeInfo.SearchAttributes == nil {
-		exeInfo.SearchAttributes = make(map[string][]byte)
-	}
-	exeInfo.SearchAttributes[definition.BinaryChecksums] = bytes
-	if common.IsAdvancedVisibilityWritingEnabled(e.shard.GetConfig().AdvancedVisibilityWritingMode(), e.shard.GetConfig().IsAdvancedVisConfigExist) {
-		return e.taskGenerator.GenerateWorkflowSearchAttrTasks()
-	}
-	return nil
-}
-
-// TODO: we will release the restriction when reset API allow those pending
-func (e *mutableStateBuilder) CheckResettable() error {
-	if len(e.GetPendingChildExecutionInfos()) > 0 {
-		return &types.BadRequestError{
-			Message: "it is not allowed resetting to a point that workflow has pending child types.",
-		}
-	}
-	if len(e.GetPendingRequestCancelExternalInfos()) > 0 {
-		return &types.BadRequestError{
-			Message: "it is not allowed resetting to a point that workflow has pending request cancel.",
-		}
-	}
-	if len(e.GetPendingSignalExternalInfos()) > 0 {
-		return &types.BadRequestError{
-			Message: "it is not allowed resetting to a point that workflow has pending signals to send.",
-		}
-	}
-	return nil
-}
-
-func (e *mutableStateBuilder) AddDecisionTaskCompletedEvent(
-	scheduleEventID int64,
-	startedEventID int64,
-	request *types.RespondDecisionTaskCompletedRequest,
-	maxResetPoints int,
-) (*types.HistoryEvent, error) {
-	opTag := tag.WorkflowActionDecisionTaskCompleted
-	if err := e.checkMutability(opTag); err != nil {
-		return nil, err
-	}
-	return e.decisionTaskManager.AddDecisionTaskCompletedEvent(scheduleEventID, startedEventID, request, maxResetPoints)
-}
-
-func (e *mutableStateBuilder) ReplicateDecisionTaskCompletedEvent(
-	event *types.HistoryEvent,
-) error {
-	return e.decisionTaskManager.ReplicateDecisionTaskCompletedEvent(event)
-}
-
-func (e *mutableStateBuilder) AddDecisionTaskTimedOutEvent(
-	scheduleEventID int64,
-	startedEventID int64,
-) (*types.HistoryEvent, error) {
-	opTag := tag.WorkflowActionDecisionTaskTimedOut
-	if err := e.checkMutability(opTag); err != nil {
-		return nil, err
-	}
-	return e.decisionTaskManager.AddDecisionTaskTimedOutEvent(scheduleEventID, startedEventID)
-}
-
-func (e *mutableStateBuilder) ReplicateDecisionTaskTimedOutEvent(
-	event *types.HistoryEvent,
-) error {
-	return e.decisionTaskManager.ReplicateDecisionTaskTimedOutEvent(event)
-}
-
-func (e *mutableStateBuilder) AddDecisionTaskScheduleToStartTimeoutEvent(
-	scheduleEventID int64,
-) (*types.HistoryEvent, error) {
-	opTag := tag.WorkflowActionDecisionTaskTimedOut
-	if err := e.checkMutability(opTag); err != nil {
-		return nil, err
-	}
-	return e.decisionTaskManager.AddDecisionTaskScheduleToStartTimeoutEvent(scheduleEventID)
-}
-
-func (e *mutableStateBuilder) AddDecisionTaskResetTimeoutEvent(
-	scheduleEventID int64,
-	baseRunID string,
-	newRunID string,
-	forkEventVersion int64,
-	reason string,
-	resetRequestID string,
-) (*types.HistoryEvent, error) {
-	opTag := tag.WorkflowActionDecisionTaskTimedOut
-	if err := e.checkMutability(opTag); err != nil {
-		return nil, err
-	}
-	return e.decisionTaskManager.AddDecisionTaskResetTimeoutEvent(
-		scheduleEventID,
-		baseRunID,
-		newRunID,
-		forkEventVersion,
-		reason,
-		resetRequestID,
-	)
-}
-
-func (e *mutableStateBuilder) AddDecisionTaskFailedEvent(
-	scheduleEventID int64,
-	startedEventID int64,
-	cause types.DecisionTaskFailedCause,
-	details []byte,
-	identity string,
-	reason string,
-	binChecksum string,
-	baseRunID string,
-	newRunID string,
-	forkEventVersion int64,
-	resetRequestID string,
-) (*types.HistoryEvent, error) {
-	opTag := tag.WorkflowActionDecisionTaskFailed
-	if err := e.checkMutability(opTag); err != nil {
-		return nil, err
-	}
-	return e.decisionTaskManager.AddDecisionTaskFailedEvent(
-		scheduleEventID,
-		startedEventID,
-		cause,
-		details,
-		identity,
-		reason,
-		binChecksum,
-		baseRunID,
-		newRunID,
-		forkEventVersion,
-		resetRequestID,
-	)
-}
-
-func (e *mutableStateBuilder) ReplicateDecisionTaskFailedEvent(event *types.HistoryEvent) error {
-	return e.decisionTaskManager.ReplicateDecisionTaskFailedEvent(event)
-}
-
-func (e *mutableStateBuilder) AddActivityTaskScheduledEvent(
-	ctx context.Context,
-	decisionCompletedEventID int64,
-	attributes *types.ScheduleActivityTaskDecisionAttributes,
-	dispatch bool,
-) (*types.HistoryEvent, *persistence.ActivityInfo, *types.ActivityLocalDispatchInfo, bool, bool, error) {
-
-	opTag := tag.WorkflowActionActivityTaskScheduled
-	if err := e.checkMutability(opTag); err != nil {
-		return nil, nil, nil, false, false, err
-	}
-
-	_, ok := e.GetActivityByActivityID(attributes.GetActivityID())
-	if ok {
-		e.logger.Warn(mutableStateInvalidHistoryActionMsg, opTag,
-			tag.WorkflowEventID(e.GetNextEventID()),
-			tag.ErrorTypeInvalidHistoryAction)
-		return nil, nil, nil, false, false, e.createCallerError(opTag)
-	}
-
-	pendingActivitiesCount := len(e.pendingActivityInfoIDs)
-
-	if pendingActivitiesCount >= e.config.PendingActivitiesCountLimitError() {
-		e.logger.Error("Pending activity count exceeds error limit",
-			tag.WorkflowDomainName(e.GetDomainEntry().GetInfo().Name),
-			tag.WorkflowID(e.executionInfo.WorkflowID),
-			tag.WorkflowRunID(e.executionInfo.RunID),
-			tag.Number(int64(pendingActivitiesCount)))
-
-		if e.config.PendingActivityValidationEnabled() {
-			return nil, nil, nil, false, false, ErrTooManyPendingActivities
-		}
-	} else if pendingActivitiesCount >= e.config.PendingActivitiesCountLimitWarn() && !e.pendingActivityWarningSent {
-		e.logger.Warn("Pending activity count exceeds warn limit",
-			tag.WorkflowDomainName(e.GetDomainEntry().GetInfo().Name),
-			tag.WorkflowID(e.executionInfo.WorkflowID),
-			tag.WorkflowRunID(e.executionInfo.RunID),
-			tag.Number(int64(pendingActivitiesCount)))
-
-		e.pendingActivityWarningSent = true
-	}
-
-	event := e.hBuilder.AddActivityTaskScheduledEvent(decisionCompletedEventID, attributes)
-
-	// Write the event to cache only on active cluster for processing on activity started or retried
-	e.eventsCache.PutEvent(
-		e.executionInfo.DomainID,
-		e.executionInfo.WorkflowID,
-		e.executionInfo.RunID,
-		event.ID,
-		event,
-	)
-
-	ai, err := e.ReplicateActivityTaskScheduledEvent(decisionCompletedEventID, event, true)
-	if err != nil {
-		return nil, nil, nil, false, false, err
-	}
-	activityStartedScope := e.metricsClient.Scope(metrics.HistoryRecordActivityTaskStartedScope)
-	if e.config.EnableActivityLocalDispatchByDomain(e.domainEntry.GetInfo().Name) && attributes.RequestLocalDispatch {
-		activityStartedScope.IncCounter(metrics.CadenceRequests)
-		return event, ai, &types.ActivityLocalDispatchInfo{ActivityID: ai.ActivityID}, false, false, nil
-	}
-	started := false
-	if dispatch {
-		started = e.tryDispatchActivityTask(ctx, event, ai)
-	}
-	if started {
-		activityStartedScope.IncCounter(metrics.CadenceRequests)
-		return event, ai, nil, true, true, nil
-	}
-
-	if err := e.taskGenerator.GenerateActivityTransferTasks(event); err != nil {
-		return nil, nil, nil, dispatch, false, err
-	}
-
-	return event, ai, nil, dispatch, false, err
-}
-
-func (e *mutableStateBuilder) tryDispatchActivityTask(
-	ctx context.Context,
-	scheduledEvent *types.HistoryEvent,
-	ai *persistence.ActivityInfo,
-) bool {
-	taggedScope := e.metricsClient.Scope(metrics.HistoryScheduleDecisionTaskScope).Tagged(
-		metrics.DomainTag(e.domainEntry.GetInfo().Name),
-		metrics.WorkflowTypeTag(e.GetWorkflowType().Name),
-		metrics.TaskListTag(ai.TaskList))
-	taggedScope.IncCounter(metrics.DecisionTypeScheduleActivityDispatchCounter)
-	err := e.shard.GetService().GetMatchingClient().AddActivityTask(ctx, &types.AddActivityTaskRequest{
-		DomainUUID:       e.executionInfo.DomainID,
-		SourceDomainUUID: e.domainEntry.GetInfo().ID,
-		Execution: &types.WorkflowExecution{
-			WorkflowID: e.executionInfo.WorkflowID,
-			RunID:      e.executionInfo.RunID,
-		},
-		TaskList:                      &types.TaskList{Name: ai.TaskList},
-		ScheduleID:                    scheduledEvent.ID,
-		ScheduleToStartTimeoutSeconds: common.Int32Ptr(ai.ScheduleToStartTimeout),
-		ActivityTaskDispatchInfo: &types.ActivityTaskDispatchInfo{
-			ScheduledEvent:                  scheduledEvent,
-			StartedTimestamp:                common.Int64Ptr(e.timeSource.Now().UnixNano()),
-			WorkflowType:                    e.GetWorkflowType(),
-			WorkflowDomain:                  e.GetDomainEntry().GetInfo().Name,
-			ScheduledTimestampOfThisAttempt: common.Int64Ptr(ai.ScheduledTime.UnixNano()),
-		},
-		PartitionConfig: e.executionInfo.PartitionConfig,
-	})
-	if err == nil {
-		taggedScope.IncCounter(metrics.DecisionTypeScheduleActivityDispatchSucceedCounter)
-		return true
-	}
-	return false
-}
-
-func (e *mutableStateBuilder) ReplicateActivityTaskScheduledEvent(
-	firstEventID int64,
-	event *types.HistoryEvent,
-	skipTaskGeneration bool,
-) (*persistence.ActivityInfo, error) {
-
-	attributes := event.ActivityTaskScheduledEventAttributes
-	targetDomainID := e.executionInfo.DomainID
-	if attributes.GetDomain() != "" {
-		var err error
-		targetDomainID, err = e.shard.GetDomainCache().GetDomainID(attributes.GetDomain())
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	scheduleEventID := event.ID
-	scheduleToCloseTimeout := attributes.GetScheduleToCloseTimeoutSeconds()
-
-	ai := &persistence.ActivityInfo{
-		Version:                  event.Version,
-		ScheduleID:               scheduleEventID,
-		ScheduledEventBatchID:    firstEventID,
-		ScheduledTime:            time.Unix(0, event.GetTimestamp()),
-		StartedID:                common.EmptyEventID,
-		StartedTime:              time.Time{},
-		ActivityID:               attributes.ActivityID,
-		DomainID:                 targetDomainID,
-		ScheduleToStartTimeout:   attributes.GetScheduleToStartTimeoutSeconds(),
-		ScheduleToCloseTimeout:   scheduleToCloseTimeout,
-		StartToCloseTimeout:      attributes.GetStartToCloseTimeoutSeconds(),
-		HeartbeatTimeout:         attributes.GetHeartbeatTimeoutSeconds(),
-		CancelRequested:          false,
-		CancelRequestID:          common.EmptyEventID,
-		LastHeartBeatUpdatedTime: time.Time{},
-		TimerTaskStatus:          TimerTaskStatusNone,
-		TaskList:                 attributes.TaskList.GetName(),
-		HasRetryPolicy:           attributes.RetryPolicy != nil,
-	}
-
-	if ai.HasRetryPolicy {
-		ai.InitialInterval = attributes.RetryPolicy.GetInitialIntervalInSeconds()
-		ai.BackoffCoefficient = attributes.RetryPolicy.GetBackoffCoefficient()
-		ai.MaximumInterval = attributes.RetryPolicy.GetMaximumIntervalInSeconds()
-		ai.MaximumAttempts = attributes.RetryPolicy.GetMaximumAttempts()
-		ai.NonRetriableErrors = attributes.RetryPolicy.NonRetriableErrorReasons
-		if attributes.RetryPolicy.GetExpirationIntervalInSeconds() != 0 {
-			ai.ExpirationTime = ai.ScheduledTime.Add(time.Duration(attributes.RetryPolicy.GetExpirationIntervalInSeconds()) * time.Second)
-		}
-	}
-
-	e.pendingActivityInfoIDs[scheduleEventID] = ai
-	e.pendingActivityIDToEventID[ai.ActivityID] = scheduleEventID
-	e.updateActivityInfos[ai.ScheduleID] = ai
-
-	if !skipTaskGeneration {
-		return ai, e.taskGenerator.GenerateActivityTransferTasks(event)
-	}
-
-	return ai, nil
-}
-
-func (e *mutableStateBuilder) addTransientActivityStartedEvent(
-	scheduleEventID int64,
-) error {
-
-	ai, ok := e.GetActivityInfo(scheduleEventID)
-	if !ok || ai.StartedID != common.TransientEventID {
-		return nil
-	}
-
-	// activity task was started (as transient event), we need to add it now.
-	event := e.hBuilder.AddActivityTaskStartedEvent(scheduleEventID, ai.Attempt, ai.RequestID, ai.StartedIdentity,
-		ai.LastFailureReason, ai.LastFailureDetails)
-	if !ai.StartedTime.IsZero() {
-		// overwrite started event time to the one recorded in ActivityInfo
-		event.Timestamp = common.Int64Ptr(ai.StartedTime.UnixNano())
-	}
-	return e.ReplicateActivityTaskStartedEvent(event)
-}
-
-func (e *mutableStateBuilder) AddActivityTaskStartedEvent(
-	ai *persistence.ActivityInfo,
-	scheduleEventID int64,
-	requestID string,
-	identity string,
-) (*types.HistoryEvent, error) {
-
-	opTag := tag.WorkflowActionActivityTaskStarted
-	if err := e.checkMutability(opTag); err != nil {
-		return nil, err
-	}
-
-	if !ai.HasRetryPolicy {
-		event := e.hBuilder.AddActivityTaskStartedEvent(scheduleEventID, ai.Attempt, requestID, identity,
-			ai.LastFailureReason, ai.LastFailureDetails)
-		if err := e.ReplicateActivityTaskStartedEvent(event); err != nil {
-			return nil, err
-		}
-		return event, nil
-	}
-
-	// we might need to retry, so do not append started event just yet,
-	// instead update mutable state and will record started event when activity task is closed
-	ai.Version = e.GetCurrentVersion()
-	ai.StartedID = common.TransientEventID
-	ai.RequestID = requestID
-	ai.StartedTime = e.timeSource.Now()
-	ai.LastHeartBeatUpdatedTime = ai.StartedTime
-	ai.StartedIdentity = identity
-	if err := e.UpdateActivity(ai); err != nil {
-		return nil, err
-	}
-	e.syncActivityTasks[ai.ScheduleID] = struct{}{}
-	return nil, nil
-}
-
-func (e *mutableStateBuilder) ReplicateActivityTaskStartedEvent(
-	event *types.HistoryEvent,
-) error {
-
-	attributes := event.ActivityTaskStartedEventAttributes
-	scheduleID := attributes.GetScheduledEventID()
-	ai, ok := e.GetActivityInfo(scheduleID)
-	if !ok {
-		e.logError(
-			fmt.Sprintf("unable to find activity event id: %v in mutable state", scheduleID),
-			tag.ErrorTypeInvalidMutableStateAction,
-		)
-		return ErrMissingActivityInfo
-	}
-
-	ai.Version = event.Version
-	ai.StartedID = event.ID
-	ai.RequestID = attributes.GetRequestID()
-	ai.StartedTime = time.Unix(0, event.GetTimestamp())
-	ai.LastHeartBeatUpdatedTime = ai.StartedTime
-	e.updateActivityInfos[ai.ScheduleID] = ai
-	return nil
-}
-
-func (e *mutableStateBuilder) AddActivityTaskCompletedEvent(
-	scheduleEventID int64,
-	startedEventID int64,
-	request *types.RespondActivityTaskCompletedRequest,
-) (*types.HistoryEvent, error) {
-
-	opTag := tag.WorkflowActionActivityTaskCompleted
-	if err := e.checkMutability(opTag); err != nil {
-		return nil, err
-	}
-
-	if ai, ok := e.GetActivityInfo(scheduleEventID); !ok || ai.StartedID != startedEventID {
-		e.logger.Warn(
-			mutableStateInvalidHistoryActionMsg,
-			opTag,
-			tag.WorkflowEventID(e.GetNextEventID()),
-			tag.ErrorTypeInvalidHistoryAction,
-			tag.Bool(ok),
-			tag.WorkflowScheduleID(scheduleEventID),
-			tag.WorkflowStartedID(startedEventID))
-		return nil, e.createInternalServerError(opTag)
-	}
-
-	if err := e.addTransientActivityStartedEvent(scheduleEventID); err != nil {
-		return nil, err
-	}
-	event := e.hBuilder.AddActivityTaskCompletedEvent(scheduleEventID, startedEventID, request)
-	if err := e.ReplicateActivityTaskCompletedEvent(event); err != nil {
-		return nil, err
-	}
-
-	return event, nil
-}
-
-func (e *mutableStateBuilder) ReplicateActivityTaskCompletedEvent(
-	event *types.HistoryEvent,
-) error {
-
-	attributes := event.ActivityTaskCompletedEventAttributes
-	scheduleID := attributes.GetScheduledEventID()
-
-	return e.DeleteActivity(scheduleID)
-}
-
-func (e *mutableStateBuilder) AddActivityTaskFailedEvent(
-	scheduleEventID int64,
-	startedEventID int64,
-	request *types.RespondActivityTaskFailedRequest,
-) (*types.HistoryEvent, error) {
-
-	opTag := tag.WorkflowActionActivityTaskFailed
-	if err := e.checkMutability(opTag); err != nil {
-		return nil, err
-	}
-
-	if ai, ok := e.GetActivityInfo(scheduleEventID); !ok || ai.StartedID != startedEventID {
-		e.logger.Warn(mutableStateInvalidHistoryActionMsg, opTag,
-			tag.WorkflowEventID(e.GetNextEventID()),
-			tag.ErrorTypeInvalidHistoryAction,
-			tag.Bool(ok),
-			tag.WorkflowScheduleID(scheduleEventID),
-			tag.WorkflowStartedID(startedEventID))
-		return nil, e.createInternalServerError(opTag)
-	}
-
-	if err := e.addTransientActivityStartedEvent(scheduleEventID); err != nil {
-		return nil, err
-	}
-	event := e.hBuilder.AddActivityTaskFailedEvent(scheduleEventID, startedEventID, request)
-	if err := e.ReplicateActivityTaskFailedEvent(event); err != nil {
-		return nil, err
-	}
-
-	return event, nil
-}
-
-func (e *mutableStateBuilder) ReplicateActivityTaskFailedEvent(
-	event *types.HistoryEvent,
-) error {
-
-	attributes := event.ActivityTaskFailedEventAttributes
-	scheduleID := attributes.GetScheduledEventID()
-
-	return e.DeleteActivity(scheduleID)
-}
-
-func (e *mutableStateBuilder) AddActivityTaskTimedOutEvent(
-	scheduleEventID int64,
-	startedEventID int64,
-	timeoutType types.TimeoutType,
-	lastHeartBeatDetails []byte,
-) (*types.HistoryEvent, error) {
-
-	opTag := tag.WorkflowActionActivityTaskTimedOut
-	if err := e.checkMutability(opTag); err != nil {
-		return nil, err
-	}
-
-	ai, ok := e.GetActivityInfo(scheduleEventID)
-	if !ok || ai.StartedID != startedEventID || ((timeoutType == types.TimeoutTypeStartToClose ||
-		timeoutType == types.TimeoutTypeHeartbeat) && ai.StartedID == common.EmptyEventID) {
-		e.logger.Warn(mutableStateInvalidHistoryActionMsg, opTag,
-			tag.WorkflowEventID(e.GetNextEventID()),
-			tag.ErrorTypeInvalidHistoryAction,
-			tag.Bool(ok),
-			tag.WorkflowScheduleID(ai.ScheduleID),
-			tag.WorkflowStartedID(ai.StartedID),
-			tag.WorkflowTimeoutType(int64(timeoutType)))
-		return nil, e.createInternalServerError(opTag)
-	}
-
-	if err := e.addTransientActivityStartedEvent(scheduleEventID); err != nil {
-		return nil, err
-	}
-	event := e.hBuilder.AddActivityTaskTimedOutEvent(scheduleEventID, startedEventID, timeoutType, lastHeartBeatDetails, ai.LastFailureReason, ai.LastFailureDetails)
-	if err := e.ReplicateActivityTaskTimedOutEvent(event); err != nil {
-		return nil, err
-	}
-
-	return event, nil
-}
-
-func (e *mutableStateBuilder) ReplicateActivityTaskTimedOutEvent(
-	event *types.HistoryEvent,
-) error {
-
-	attributes := event.ActivityTaskTimedOutEventAttributes
-	scheduleID := attributes.GetScheduledEventID()
-
-	return e.DeleteActivity(scheduleID)
-}
-
-func (e *mutableStateBuilder) AddActivityTaskCancelRequestedEvent(
-	decisionCompletedEventID int64,
-	activityID string,
-	identity string,
-) (*types.HistoryEvent, *persistence.ActivityInfo, error) {
-
-	opTag := tag.WorkflowActionActivityTaskCancelRequested
-	if err := e.checkMutability(opTag); err != nil {
-		return nil, nil, err
-	}
-
-	// we need to add the cancel request event even if activity not in mutable state
-	// if activity not in mutable state or already cancel requested,
-	// we do not need to call the replication function
-	actCancelReqEvent := e.hBuilder.AddActivityTaskCancelRequestedEvent(decisionCompletedEventID, activityID)
-
-	ai, ok := e.GetActivityByActivityID(activityID)
-	if !ok || ai.CancelRequested {
-		e.logWarn(mutableStateInvalidHistoryActionMsg, opTag,
-			tag.WorkflowEventID(e.GetNextEventID()),
-			tag.ErrorTypeInvalidHistoryAction,
-			tag.Bool(ok),
-			tag.WorkflowActivityID(activityID))
-
-		return nil, nil, e.createCallerError(opTag)
-	}
-
-	if err := e.ReplicateActivityTaskCancelRequestedEvent(actCancelReqEvent); err != nil {
-		return nil, nil, err
-	}
-
-	return actCancelReqEvent, ai, nil
-}
-
-func (e *mutableStateBuilder) ReplicateActivityTaskCancelRequestedEvent(
-	event *types.HistoryEvent,
-) error {
-
-	attributes := event.ActivityTaskCancelRequestedEventAttributes
-	activityID := attributes.GetActivityID()
-	ai, ok := e.GetActivityByActivityID(activityID)
-	if !ok {
-		// On active side, if the ActivityTaskCancelRequested is invalid, it will created a RequestCancelActivityTaskFailed
-		// Passive will rely on active side logic
-		return nil
-	}
-
-	ai.Version = event.Version
-
-	// - We have the activity dispatched to worker.
-	// - The activity might not be heartbeating, but the activity can still call RecordActivityHeartBeat()
-	//   to see cancellation while reporting progress of the activity.
-	ai.CancelRequested = true
-
-	ai.CancelRequestID = event.ID
-	e.updateActivityInfos[ai.ScheduleID] = ai
-	return nil
-}
-
-func (e *mutableStateBuilder) AddRequestCancelActivityTaskFailedEvent(
-	decisionCompletedEventID int64,
-	activityID string,
-	cause string,
-) (*types.HistoryEvent, error) {
-
-	opTag := tag.WorkflowActionActivityTaskCancelFailed
-	if err := e.checkMutability(opTag); err != nil {
-		return nil, err
-	}
-
-	return e.hBuilder.AddRequestCancelActivityTaskFailedEvent(decisionCompletedEventID, activityID, cause), nil
-}
-
-func (e *mutableStateBuilder) AddActivityTaskCanceledEvent(
-	scheduleEventID int64,
-	startedEventID int64,
-	latestCancelRequestedEventID int64,
-	details []byte,
-	identity string,
-) (*types.HistoryEvent, error) {
-
-	opTag := tag.WorkflowActionActivityTaskCanceled
-	if err := e.checkMutability(opTag); err != nil {
-		return nil, err
-	}
-
-	ai, ok := e.GetActivityInfo(scheduleEventID)
-	if !ok || ai.StartedID != startedEventID {
-		e.logWarn(mutableStateInvalidHistoryActionMsg, opTag,
-			tag.WorkflowEventID(e.GetNextEventID()),
-			tag.ErrorTypeInvalidHistoryAction,
-			tag.WorkflowScheduleID(scheduleEventID))
-		return nil, e.createInternalServerError(opTag)
-	}
-
-	// Verify cancel request as well.
-	if !ai.CancelRequested {
-		e.logWarn(mutableStateInvalidHistoryActionMsg, opTag,
-			tag.WorkflowEventID(e.GetNextEventID()),
-			tag.ErrorTypeInvalidHistoryAction,
-			tag.WorkflowScheduleID(scheduleEventID),
-			tag.WorkflowActivityID(ai.ActivityID),
-			tag.WorkflowStartedID(ai.StartedID))
-		return nil, e.createInternalServerError(opTag)
-	}
-
-	if err := e.addTransientActivityStartedEvent(scheduleEventID); err != nil {
-		return nil, err
-	}
-	event := e.hBuilder.AddActivityTaskCanceledEvent(scheduleEventID, startedEventID, latestCancelRequestedEventID,
-		details, identity)
-	if err := e.ReplicateActivityTaskCanceledEvent(event); err != nil {
-		return nil, err
-	}
-
-	return event, nil
-}
-
-func (e *mutableStateBuilder) ReplicateActivityTaskCanceledEvent(
-	event *types.HistoryEvent,
-) error {
-
-	attributes := event.ActivityTaskCanceledEventAttributes
-	scheduleID := attributes.GetScheduledEventID()
-
-	return e.DeleteActivity(scheduleID)
-}
-
-func (e *mutableStateBuilder) AddCompletedWorkflowEvent(
-	decisionCompletedEventID int64,
-	attributes *types.CompleteWorkflowExecutionDecisionAttributes,
-) (*types.HistoryEvent, error) {
-
-	opTag := tag.WorkflowActionWorkflowCompleted
-	if err := e.checkMutability(opTag); err != nil {
-		return nil, err
-	}
-
-	event := e.hBuilder.AddCompletedWorkflowEvent(decisionCompletedEventID, attributes)
-	if err := e.ReplicateWorkflowExecutionCompletedEvent(decisionCompletedEventID, event); err != nil {
-		return nil, err
-	}
-	return event, nil
-}
-
-func (e *mutableStateBuilder) ReplicateWorkflowExecutionCompletedEvent(
-	firstEventID int64,
-	event *types.HistoryEvent,
-) error {
-
-	if err := e.UpdateWorkflowStateCloseStatus(
-		persistence.WorkflowStateCompleted,
-		persistence.WorkflowCloseStatusCompleted,
-	); err != nil {
-		return err
-	}
-	e.executionInfo.CompletionEventBatchID = firstEventID // Used when completion event needs to be loaded from database
-	e.ClearStickyness()
-	e.writeEventToCache(event)
-
-	return e.taskGenerator.GenerateWorkflowCloseTasks(event, e.config.WorkflowDeletionJitterRange(e.domainEntry.GetInfo().Name))
-}
-
-func (e *mutableStateBuilder) AddFailWorkflowEvent(
-	decisionCompletedEventID int64,
-	attributes *types.FailWorkflowExecutionDecisionAttributes,
-) (*types.HistoryEvent, error) {
-
-	opTag := tag.WorkflowActionWorkflowFailed
-	if err := e.checkMutability(opTag); err != nil {
-		return nil, err
-	}
-
-	event := e.hBuilder.AddFailWorkflowEvent(decisionCompletedEventID, attributes)
-	if err := e.ReplicateWorkflowExecutionFailedEvent(decisionCompletedEventID, event); err != nil {
-		return nil, err
-	}
-	return event, nil
-}
-
-func (e *mutableStateBuilder) ReplicateWorkflowExecutionFailedEvent(
-	firstEventID int64,
-	event *types.HistoryEvent,
-) error {
-
-	if err := e.UpdateWorkflowStateCloseStatus(
-		persistence.WorkflowStateCompleted,
-		persistence.WorkflowCloseStatusFailed,
-	); err != nil {
-		return err
-	}
-	e.executionInfo.CompletionEventBatchID = firstEventID // Used when completion event needs to be loaded from database
-	e.ClearStickyness()
-	e.writeEventToCache(event)
-
-	return e.taskGenerator.GenerateWorkflowCloseTasks(event, e.config.WorkflowDeletionJitterRange(e.domainEntry.GetInfo().Name))
-}
-
-func (e *mutableStateBuilder) AddTimeoutWorkflowEvent(
-	firstEventID int64,
-) (*types.HistoryEvent, error) {
-
-	opTag := tag.WorkflowActionWorkflowTimeout
-	if err := e.checkMutability(opTag); err != nil {
-		return nil, err
-	}
-
-	event := e.hBuilder.AddTimeoutWorkflowEvent()
-	if err := e.ReplicateWorkflowExecutionTimedoutEvent(firstEventID, event); err != nil {
-		return nil, err
-	}
-	return event, nil
-}
-
-func (e *mutableStateBuilder) ReplicateWorkflowExecutionTimedoutEvent(
-	firstEventID int64,
-	event *types.HistoryEvent,
-) error {
-
-	if err := e.UpdateWorkflowStateCloseStatus(
-		persistence.WorkflowStateCompleted,
-		persistence.WorkflowCloseStatusTimedOut,
-	); err != nil {
-		return err
-	}
-	e.executionInfo.CompletionEventBatchID = firstEventID // Used when completion event needs to be loaded from database
-	e.ClearStickyness()
-	e.writeEventToCache(event)
-
-	return e.taskGenerator.GenerateWorkflowCloseTasks(event, e.config.WorkflowDeletionJitterRange(e.domainEntry.GetInfo().Name))
-}
-
-func (e *mutableStateBuilder) AddWorkflowExecutionCancelRequestedEvent(
-	cause string,
-	request *types.HistoryRequestCancelWorkflowExecutionRequest,
-) (*types.HistoryEvent, error) {
-
-	opTag := tag.WorkflowActionWorkflowCancelRequested
-	if err := e.checkMutability(opTag); err != nil {
-		return nil, err
-	}
-
-	if e.executionInfo.CancelRequested {
-		e.logWarn(mutableStateInvalidHistoryActionMsg, opTag,
-			tag.WorkflowEventID(e.GetNextEventID()),
-			tag.ErrorTypeInvalidHistoryAction,
-			tag.WorkflowState(e.executionInfo.State),
-			tag.Bool(e.executionInfo.CancelRequested),
-			tag.Key(e.executionInfo.CancelRequestID),
-		)
-		return nil, e.createInternalServerError(opTag)
-	}
-
-	event := e.hBuilder.AddWorkflowExecutionCancelRequestedEvent(cause, request)
-	if err := e.ReplicateWorkflowExecutionCancelRequestedEvent(event); err != nil {
-		return nil, err
-	}
-	return event, nil
-}
-
-func (e *mutableStateBuilder) ReplicateWorkflowExecutionCancelRequestedEvent(
-	event *types.HistoryEvent,
-) error {
-
-	e.executionInfo.CancelRequested = true
-	requestID := event.WorkflowExecutionCancelRequestedEventAttributes.RequestID
-	e.executionInfo.CancelRequestID = requestID
-	e.insertWorkflowRequest(persistence.WorkflowRequest{
-		RequestID:   requestID,
-		Version:     event.Version,
-		RequestType: persistence.WorkflowRequestTypeCancel,
-	})
-	return nil
-}
-
-func (e *mutableStateBuilder) AddWorkflowExecutionCanceledEvent(
-	decisionTaskCompletedEventID int64,
-	attributes *types.CancelWorkflowExecutionDecisionAttributes,
-) (*types.HistoryEvent, error) {
-
-	opTag := tag.WorkflowActionWorkflowCanceled
-	if err := e.checkMutability(opTag); err != nil {
-		return nil, err
-	}
-
-	event := e.hBuilder.AddWorkflowExecutionCanceledEvent(decisionTaskCompletedEventID, attributes)
-	if err := e.ReplicateWorkflowExecutionCanceledEvent(decisionTaskCompletedEventID, event); err != nil {
-		return nil, err
-	}
-	return event, nil
-}
-
-func (e *mutableStateBuilder) ReplicateWorkflowExecutionCanceledEvent(
-	firstEventID int64,
-	event *types.HistoryEvent,
-) error {
-	if err := e.UpdateWorkflowStateCloseStatus(
-		persistence.WorkflowStateCompleted,
-		persistence.WorkflowCloseStatusCanceled,
-	); err != nil {
-		return err
-	}
-	e.executionInfo.CompletionEventBatchID = firstEventID // Used when completion event needs to be loaded from database
-	e.ClearStickyness()
-	e.writeEventToCache(event)
-
-	return e.taskGenerator.GenerateWorkflowCloseTasks(event, e.config.WorkflowDeletionJitterRange(e.domainEntry.GetInfo().Name))
-}
-
-func (e *mutableStateBuilder) AddRequestCancelExternalWorkflowExecutionInitiatedEvent(
-	decisionCompletedEventID int64,
-	cancelRequestID string,
-	request *types.RequestCancelExternalWorkflowExecutionDecisionAttributes,
-) (*types.HistoryEvent, *persistence.RequestCancelInfo, error) {
-
-	opTag := tag.WorkflowActionExternalWorkflowCancelInitiated
-	if err := e.checkMutability(opTag); err != nil {
-		return nil, nil, err
-	}
-
-	event := e.hBuilder.AddRequestCancelExternalWorkflowExecutionInitiatedEvent(decisionCompletedEventID, request)
-	rci, err := e.ReplicateRequestCancelExternalWorkflowExecutionInitiatedEvent(decisionCompletedEventID, event, cancelRequestID)
-	if err != nil {
-		return nil, nil, err
-	}
-	return event, rci, nil
-}
-
-func (e *mutableStateBuilder) ReplicateRequestCancelExternalWorkflowExecutionInitiatedEvent(
-	firstEventID int64,
-	event *types.HistoryEvent,
-	cancelRequestID string,
-) (*persistence.RequestCancelInfo, error) {
-
-	// TODO: Evaluate if we need cancelRequestID also part of history event
-	initiatedEventID := event.ID
-	rci := &persistence.RequestCancelInfo{
-		Version:               event.Version,
-		InitiatedEventBatchID: firstEventID,
-		InitiatedID:           initiatedEventID,
-		CancelRequestID:       cancelRequestID,
-	}
-
-	e.pendingRequestCancelInfoIDs[rci.InitiatedID] = rci
-	e.updateRequestCancelInfos[rci.InitiatedID] = rci
-
-	return rci, e.taskGenerator.GenerateRequestCancelExternalTasks(event)
-}
-
-func (e *mutableStateBuilder) AddExternalWorkflowExecutionCancelRequested(
-	initiatedID int64,
-	domain string,
-	workflowID string,
-	runID string,
-) (*types.HistoryEvent, error) {
-
-	opTag := tag.WorkflowActionExternalWorkflowCancelRequested
-	if err := e.checkMutability(opTag); err != nil {
-		return nil, err
-	}
-
-	_, ok := e.GetRequestCancelInfo(initiatedID)
-	if !ok {
-		e.logWarn(mutableStateInvalidHistoryActionMsg, opTag,
-			tag.WorkflowEventID(e.GetNextEventID()),
-			tag.ErrorTypeInvalidHistoryAction,
-			tag.WorkflowInitiatedID(initiatedID))
-		return nil, e.createInternalServerError(opTag)
-	}
-
-	event := e.hBuilder.AddExternalWorkflowExecutionCancelRequested(initiatedID, domain, workflowID, runID)
-	if err := e.ReplicateExternalWorkflowExecutionCancelRequested(event); err != nil {
-		return nil, err
-	}
-	return event, nil
-}
-
-func (e *mutableStateBuilder) ReplicateExternalWorkflowExecutionCancelRequested(
-	event *types.HistoryEvent,
-) error {
-
-	initiatedID := event.ExternalWorkflowExecutionCancelRequestedEventAttributes.GetInitiatedEventID()
-
-	return e.DeletePendingRequestCancel(initiatedID)
-}
-
-func (e *mutableStateBuilder) AddRequestCancelExternalWorkflowExecutionFailedEvent(
-	decisionTaskCompletedEventID int64,
-	initiatedID int64,
-	domain string,
-	workflowID string,
-	runID string,
-	cause types.CancelExternalWorkflowExecutionFailedCause,
-) (*types.HistoryEvent, error) {
-
-	opTag := tag.WorkflowActionExternalWorkflowCancelFailed
-	if err := e.checkMutability(opTag); err != nil {
-		return nil, err
-	}
-
-	_, ok := e.GetRequestCancelInfo(initiatedID)
-	if !ok {
-		e.logWarn(mutableStateInvalidHistoryActionMsg, opTag,
-			tag.WorkflowEventID(e.GetNextEventID()),
-			tag.ErrorTypeInvalidHistoryAction,
-			tag.WorkflowInitiatedID(initiatedID))
-		return nil, e.createInternalServerError(opTag)
-	}
-
-	event := e.hBuilder.AddRequestCancelExternalWorkflowExecutionFailedEvent(decisionTaskCompletedEventID, initiatedID,
-		domain, workflowID, runID, cause)
-	if err := e.ReplicateRequestCancelExternalWorkflowExecutionFailedEvent(event); err != nil {
-		return nil, err
-	}
-	return event, nil
-}
-
-func (e *mutableStateBuilder) ReplicateRequestCancelExternalWorkflowExecutionFailedEvent(
-	event *types.HistoryEvent,
-) error {
-
-	initiatedID := event.RequestCancelExternalWorkflowExecutionFailedEventAttributes.GetInitiatedEventID()
-
-	return e.DeletePendingRequestCancel(initiatedID)
-}
-
-func (e *mutableStateBuilder) AddSignalExternalWorkflowExecutionInitiatedEvent(
-	decisionCompletedEventID int64,
-	signalRequestID string,
-	request *types.SignalExternalWorkflowExecutionDecisionAttributes,
-) (*types.HistoryEvent, *persistence.SignalInfo, error) {
-
-	opTag := tag.WorkflowActionExternalWorkflowSignalInitiated
-	if err := e.checkMutability(opTag); err != nil {
-		return nil, nil, err
-	}
-
-	event := e.hBuilder.AddSignalExternalWorkflowExecutionInitiatedEvent(decisionCompletedEventID, request)
-	si, err := e.ReplicateSignalExternalWorkflowExecutionInitiatedEvent(decisionCompletedEventID, event, signalRequestID)
-	if err != nil {
-		return nil, nil, err
-	}
-	return event, si, nil
-}
-
-func (e *mutableStateBuilder) ReplicateSignalExternalWorkflowExecutionInitiatedEvent(
-	firstEventID int64,
-	event *types.HistoryEvent,
-	signalRequestID string,
-) (*persistence.SignalInfo, error) {
-
-	// TODO: Consider also writing signalRequestID to history event
-	initiatedEventID := event.ID
-	attributes := event.SignalExternalWorkflowExecutionInitiatedEventAttributes
-	si := &persistence.SignalInfo{
-		Version:               event.Version,
-		InitiatedEventBatchID: firstEventID,
-		InitiatedID:           initiatedEventID,
-		SignalRequestID:       signalRequestID,
-		SignalName:            attributes.GetSignalName(),
-		Input:                 attributes.Input,
-		Control:               attributes.Control,
-	}
-
-	e.pendingSignalInfoIDs[si.InitiatedID] = si
-	e.updateSignalInfos[si.InitiatedID] = si
-
-	return si, e.taskGenerator.GenerateSignalExternalTasks(event)
-}
-
 func (e *mutableStateBuilder) AddUpsertWorkflowSearchAttributesEvent(
 	decisionCompletedEventID int64,
 	request *types.UpsertWorkflowSearchAttributesDecisionAttributes,
@@ -3054,257 +1148,6 @@ func (e *mutableStateBuilder) ReplicateUpsertWorkflowSearchAttributesEvent(
 	e.executionInfo.SearchAttributes = mergeMapOfByteArray(currentSearchAttr, upsertSearchAttr)
 
 	return e.taskGenerator.GenerateWorkflowSearchAttrTasks()
-}
-
-func mergeMapOfByteArray(
-	current map[string][]byte,
-	upsert map[string][]byte,
-) map[string][]byte {
-
-	if current == nil {
-		current = make(map[string][]byte)
-	}
-	for k, v := range upsert {
-		current[k] = v
-	}
-	return current
-}
-
-func (e *mutableStateBuilder) AddExternalWorkflowExecutionSignaled(
-	initiatedID int64,
-	domain string,
-	workflowID string,
-	runID string,
-	control []byte,
-) (*types.HistoryEvent, error) {
-
-	opTag := tag.WorkflowActionExternalWorkflowSignalRequested
-	if err := e.checkMutability(opTag); err != nil {
-		return nil, err
-	}
-
-	_, ok := e.GetSignalInfo(initiatedID)
-	if !ok {
-		e.logWarn(mutableStateInvalidHistoryActionMsg, opTag,
-			tag.WorkflowEventID(e.GetNextEventID()),
-			tag.ErrorTypeInvalidHistoryAction,
-			tag.WorkflowInitiatedID(initiatedID))
-		return nil, e.createInternalServerError(opTag)
-	}
-
-	event := e.hBuilder.AddExternalWorkflowExecutionSignaled(initiatedID, domain, workflowID, runID, control)
-	if err := e.ReplicateExternalWorkflowExecutionSignaled(event); err != nil {
-		return nil, err
-	}
-	return event, nil
-}
-
-func (e *mutableStateBuilder) ReplicateExternalWorkflowExecutionSignaled(
-	event *types.HistoryEvent,
-) error {
-
-	initiatedID := event.ExternalWorkflowExecutionSignaledEventAttributes.GetInitiatedEventID()
-
-	return e.DeletePendingSignal(initiatedID)
-}
-
-func (e *mutableStateBuilder) AddSignalExternalWorkflowExecutionFailedEvent(
-	decisionTaskCompletedEventID int64,
-	initiatedID int64,
-	domain string,
-	workflowID string,
-	runID string,
-	control []byte,
-	cause types.SignalExternalWorkflowExecutionFailedCause,
-) (*types.HistoryEvent, error) {
-
-	opTag := tag.WorkflowActionExternalWorkflowSignalFailed
-	if err := e.checkMutability(opTag); err != nil {
-		return nil, err
-	}
-
-	_, ok := e.GetSignalInfo(initiatedID)
-	if !ok {
-		e.logWarn(mutableStateInvalidHistoryActionMsg, opTag,
-			tag.WorkflowEventID(e.GetNextEventID()),
-			tag.ErrorTypeInvalidHistoryAction,
-			tag.WorkflowInitiatedID(initiatedID))
-		return nil, e.createInternalServerError(opTag)
-	}
-
-	event := e.hBuilder.AddSignalExternalWorkflowExecutionFailedEvent(decisionTaskCompletedEventID, initiatedID, domain,
-		workflowID, runID, control, cause)
-	if err := e.ReplicateSignalExternalWorkflowExecutionFailedEvent(event); err != nil {
-		return nil, err
-	}
-	return event, nil
-}
-
-func (e *mutableStateBuilder) ReplicateSignalExternalWorkflowExecutionFailedEvent(
-	event *types.HistoryEvent,
-) error {
-
-	initiatedID := event.SignalExternalWorkflowExecutionFailedEventAttributes.GetInitiatedEventID()
-
-	return e.DeletePendingSignal(initiatedID)
-}
-
-func (e *mutableStateBuilder) AddTimerStartedEvent(
-	decisionCompletedEventID int64,
-	request *types.StartTimerDecisionAttributes,
-) (*types.HistoryEvent, *persistence.TimerInfo, error) {
-
-	opTag := tag.WorkflowActionTimerStarted
-	if err := e.checkMutability(opTag); err != nil {
-		return nil, nil, err
-	}
-
-	timerID := request.GetTimerID()
-	_, ok := e.GetUserTimerInfo(timerID)
-	if ok {
-		e.logWarn(mutableStateInvalidHistoryActionMsg, opTag,
-			tag.WorkflowEventID(e.GetNextEventID()),
-			tag.ErrorTypeInvalidHistoryAction,
-			tag.WorkflowTimerID(timerID))
-		return nil, nil, e.createCallerError(opTag)
-	}
-
-	event := e.hBuilder.AddTimerStartedEvent(decisionCompletedEventID, request)
-	ti, err := e.ReplicateTimerStartedEvent(event)
-	if err != nil {
-		return nil, nil, err
-	}
-	return event, ti, err
-}
-
-func (e *mutableStateBuilder) ReplicateTimerStartedEvent(
-	event *types.HistoryEvent,
-) (*persistence.TimerInfo, error) {
-
-	attributes := event.TimerStartedEventAttributes
-	timerID := attributes.GetTimerID()
-
-	startToFireTimeout := attributes.GetStartToFireTimeoutSeconds()
-	fireTimeout := time.Duration(startToFireTimeout) * time.Second
-	// TODO: Time skew need to be taken in to account.
-	expiryTime := time.Unix(0, event.GetTimestamp()).Add(fireTimeout) // should use the event time, not now
-	ti := &persistence.TimerInfo{
-		Version:    event.Version,
-		TimerID:    timerID,
-		ExpiryTime: expiryTime,
-		StartedID:  event.ID,
-		TaskStatus: TimerTaskStatusNone,
-	}
-
-	e.pendingTimerInfoIDs[ti.TimerID] = ti
-	e.pendingTimerEventIDToID[ti.StartedID] = ti.TimerID
-	e.updateTimerInfos[ti.TimerID] = ti
-
-	return ti, nil
-}
-
-func (e *mutableStateBuilder) AddTimerFiredEvent(
-	timerID string,
-) (*types.HistoryEvent, error) {
-
-	opTag := tag.WorkflowActionTimerFired
-	if err := e.checkMutability(opTag); err != nil {
-		return nil, err
-	}
-
-	timerInfo, ok := e.GetUserTimerInfo(timerID)
-	if !ok {
-		e.logWarn(mutableStateInvalidHistoryActionMsg, opTag,
-			tag.WorkflowEventID(e.GetNextEventID()),
-			tag.ErrorTypeInvalidHistoryAction,
-			tag.WorkflowTimerID(timerID))
-		return nil, e.createInternalServerError(opTag)
-	}
-
-	// Timer is running.
-	event := e.hBuilder.AddTimerFiredEvent(timerInfo.StartedID, timerID)
-	if err := e.ReplicateTimerFiredEvent(event); err != nil {
-		return nil, err
-	}
-	return event, nil
-}
-
-func (e *mutableStateBuilder) ReplicateTimerFiredEvent(
-	event *types.HistoryEvent,
-) error {
-
-	attributes := event.TimerFiredEventAttributes
-	timerID := attributes.GetTimerID()
-
-	return e.DeleteUserTimer(timerID)
-}
-
-func (e *mutableStateBuilder) AddTimerCanceledEvent(
-	decisionCompletedEventID int64,
-	attributes *types.CancelTimerDecisionAttributes,
-	identity string,
-) (*types.HistoryEvent, error) {
-
-	opTag := tag.WorkflowActionTimerCanceled
-	if err := e.checkMutability(opTag); err != nil {
-		return nil, err
-	}
-
-	var timerStartedID int64
-	timerID := attributes.GetTimerID()
-	ti, ok := e.GetUserTimerInfo(timerID)
-	if !ok {
-		// if timer is not running then check if it has fired in the mutable state.
-		// If so clear the timer from the mutable state. We need to check both the
-		// bufferedEvents and the history builder
-		timerFiredEvent := e.checkAndClearTimerFiredEvent(timerID)
-		if timerFiredEvent == nil {
-			e.logWarn(mutableStateInvalidHistoryActionMsg, opTag,
-				tag.WorkflowEventID(e.GetNextEventID()),
-				tag.ErrorTypeInvalidHistoryAction,
-				tag.WorkflowTimerID(timerID))
-			return nil, e.createCallerError(opTag)
-		}
-		timerStartedID = timerFiredEvent.TimerFiredEventAttributes.GetStartedEventID()
-	} else {
-		timerStartedID = ti.StartedID
-	}
-
-	// Timer is running.
-	event := e.hBuilder.AddTimerCanceledEvent(timerStartedID, decisionCompletedEventID, timerID, identity)
-	if ok {
-		if err := e.ReplicateTimerCanceledEvent(event); err != nil {
-			return nil, err
-		}
-	}
-	return event, nil
-}
-
-func (e *mutableStateBuilder) ReplicateTimerCanceledEvent(
-	event *types.HistoryEvent,
-) error {
-
-	attributes := event.TimerCanceledEventAttributes
-	timerID := attributes.GetTimerID()
-
-	return e.DeleteUserTimer(timerID)
-}
-
-func (e *mutableStateBuilder) AddCancelTimerFailedEvent(
-	decisionCompletedEventID int64,
-	attributes *types.CancelTimerDecisionAttributes,
-	identity string,
-) (*types.HistoryEvent, error) {
-
-	opTag := tag.WorkflowActionTimerCancelFailed
-	if err := e.checkMutability(opTag); err != nil {
-		return nil, err
-	}
-
-	// No Operation: We couldn't cancel it probably TIMER_ID_UNKNOWN
-	timerID := attributes.GetTimerID()
-	return e.hBuilder.AddCancelTimerFailedEvent(timerID, decisionCompletedEventID,
-		timerCancellationMsgTimerIDUnknown, identity), nil
 }
 
 func (e *mutableStateBuilder) AddRecordMarkerEvent(
@@ -3355,39 +1198,6 @@ func (e *mutableStateBuilder) ReplicateWorkflowExecutionTerminatedEvent(
 	e.writeEventToCache(event)
 
 	return e.taskGenerator.GenerateWorkflowCloseTasks(event, e.config.WorkflowDeletionJitterRange(e.domainEntry.GetInfo().Name))
-}
-
-func (e *mutableStateBuilder) AddWorkflowExecutionSignaled(
-	signalName string,
-	input []byte,
-	identity string,
-	requestID string,
-) (*types.HistoryEvent, error) {
-
-	opTag := tag.WorkflowActionWorkflowSignaled
-	if err := e.checkMutability(opTag); err != nil {
-		return nil, err
-	}
-
-	event := e.hBuilder.AddWorkflowExecutionSignaledEvent(signalName, input, identity, requestID)
-	if err := e.ReplicateWorkflowExecutionSignaled(event); err != nil {
-		return nil, err
-	}
-	return event, nil
-}
-
-func (e *mutableStateBuilder) ReplicateWorkflowExecutionSignaled(
-	event *types.HistoryEvent,
-) error {
-
-	// Increment signal count in mutable state for this workflow execution
-	e.executionInfo.SignalCount++
-	e.insertWorkflowRequest(persistence.WorkflowRequest{
-		RequestID:   event.WorkflowExecutionSignaledEventAttributes.RequestID,
-		Version:     event.Version,
-		RequestType: persistence.WorkflowRequestTypeSignal,
-	})
-	return nil
 }
 
 func (e *mutableStateBuilder) AddContinueAsNewEvent(
@@ -3506,483 +1316,6 @@ func (e *mutableStateBuilder) ReplicateWorkflowExecutionContinuedAsNewEvent(
 	e.writeEventToCache(continueAsNewEvent)
 
 	return e.taskGenerator.GenerateWorkflowCloseTasks(continueAsNewEvent, e.config.WorkflowDeletionJitterRange(e.domainEntry.GetInfo().Name))
-}
-
-func (e *mutableStateBuilder) AddStartChildWorkflowExecutionInitiatedEvent(
-	decisionCompletedEventID int64,
-	createRequestID string,
-	attributes *types.StartChildWorkflowExecutionDecisionAttributes,
-) (*types.HistoryEvent, *persistence.ChildExecutionInfo, error) {
-
-	opTag := tag.WorkflowActionChildWorkflowInitiated
-	if err := e.checkMutability(opTag); err != nil {
-		return nil, nil, err
-	}
-
-	event := e.hBuilder.AddStartChildWorkflowExecutionInitiatedEvent(
-		decisionCompletedEventID,
-		attributes,
-		e.GetDomainEntry().GetInfo().Name,
-	)
-	// Write the event to cache only on active cluster
-	e.eventsCache.PutEvent(e.executionInfo.DomainID, e.executionInfo.WorkflowID, e.executionInfo.RunID,
-		event.ID, event)
-
-	ci, err := e.ReplicateStartChildWorkflowExecutionInitiatedEvent(decisionCompletedEventID, event, createRequestID)
-	if err != nil {
-		return nil, nil, err
-	}
-	return event, ci, nil
-}
-
-func (e *mutableStateBuilder) ReplicateStartChildWorkflowExecutionInitiatedEvent(
-	firstEventID int64,
-	event *types.HistoryEvent,
-	createRequestID string,
-) (*persistence.ChildExecutionInfo, error) {
-
-	initiatedEventID := event.ID
-	attributes := event.StartChildWorkflowExecutionInitiatedEventAttributes
-
-	domainID := e.GetExecutionInfo().DomainID
-	if domainName := attributes.GetDomain(); domainName != "" {
-		// domainName may still be empty if two cadence clusters are running different versions
-		var err error
-		domainID, err = e.shard.GetDomainCache().GetDomainID(domainName)
-		if err != nil {
-			return nil, err
-		}
-	}
-	ci := &persistence.ChildExecutionInfo{
-		Version:               event.Version,
-		InitiatedID:           initiatedEventID,
-		InitiatedEventBatchID: firstEventID,
-		StartedID:             common.EmptyEventID,
-		StartedWorkflowID:     attributes.GetWorkflowID(),
-		CreateRequestID:       createRequestID,
-		DomainID:              domainID,
-		// DomainName field is being deprecated
-		// DomainName:            attributes.GetDomain(),
-		WorkflowTypeName:  attributes.GetWorkflowType().GetName(),
-		ParentClosePolicy: attributes.GetParentClosePolicy(),
-	}
-
-	e.pendingChildExecutionInfoIDs[ci.InitiatedID] = ci
-	e.updateChildExecutionInfos[ci.InitiatedID] = ci
-
-	return ci, e.taskGenerator.GenerateChildWorkflowTasks(event)
-}
-
-func (e *mutableStateBuilder) AddChildWorkflowExecutionStartedEvent(
-	domain string,
-	execution *types.WorkflowExecution,
-	workflowType *types.WorkflowType,
-	initiatedID int64,
-	header *types.Header,
-) (*types.HistoryEvent, error) {
-
-	opTag := tag.WorkflowActionChildWorkflowStarted
-	if err := e.checkMutability(opTag); err != nil {
-		return nil, err
-	}
-
-	ci, ok := e.GetChildExecutionInfo(initiatedID)
-	if !ok || ci.StartedID != common.EmptyEventID {
-		e.logWarn(mutableStateInvalidHistoryActionMsg, opTag,
-			tag.WorkflowEventID(e.GetNextEventID()),
-			tag.ErrorTypeInvalidHistoryAction,
-			tag.Bool(ok),
-			tag.WorkflowInitiatedID(initiatedID))
-		return nil, e.createInternalServerError(opTag)
-	}
-
-	event := e.hBuilder.AddChildWorkflowExecutionStartedEvent(domain, execution, workflowType, initiatedID, header)
-	if err := e.ReplicateChildWorkflowExecutionStartedEvent(event); err != nil {
-		return nil, err
-	}
-	return event, nil
-}
-
-func (e *mutableStateBuilder) ReplicateChildWorkflowExecutionStartedEvent(
-	event *types.HistoryEvent,
-) error {
-
-	attributes := event.ChildWorkflowExecutionStartedEventAttributes
-	initiatedID := attributes.GetInitiatedEventID()
-
-	ci, ok := e.GetChildExecutionInfo(initiatedID)
-	if !ok {
-		e.logError(
-			"Unable to find child workflow",
-			tag.ErrorTypeInvalidMutableStateAction,
-			tag.WorkflowEventID(e.GetNextEventID()),
-			tag.WorkflowInitiatedID(initiatedID),
-		)
-		return ErrMissingChildWorkflowInfo
-	}
-	ci.StartedID = event.ID
-	ci.StartedRunID = attributes.GetWorkflowExecution().GetRunID()
-	e.updateChildExecutionInfos[ci.InitiatedID] = ci
-
-	return nil
-}
-
-func (e *mutableStateBuilder) AddStartChildWorkflowExecutionFailedEvent(
-	initiatedID int64,
-	cause types.ChildWorkflowExecutionFailedCause,
-	initiatedEventAttributes *types.StartChildWorkflowExecutionInitiatedEventAttributes,
-) (*types.HistoryEvent, error) {
-
-	opTag := tag.WorkflowActionChildWorkflowInitiationFailed
-	if err := e.checkMutability(opTag); err != nil {
-		return nil, err
-	}
-
-	ci, ok := e.GetChildExecutionInfo(initiatedID)
-	if !ok || ci.StartedID != common.EmptyEventID {
-		e.logWarn(mutableStateInvalidHistoryActionMsg, opTag,
-			tag.WorkflowEventID(e.GetNextEventID()),
-			tag.ErrorTypeInvalidHistoryAction,
-			tag.Bool(ok),
-			tag.WorkflowInitiatedID(initiatedID))
-		return nil, e.createInternalServerError(opTag)
-	}
-
-	event := e.hBuilder.AddStartChildWorkflowExecutionFailedEvent(initiatedID, cause, initiatedEventAttributes)
-	if err := e.ReplicateStartChildWorkflowExecutionFailedEvent(event); err != nil {
-		return nil, err
-	}
-	return event, nil
-}
-
-func (e *mutableStateBuilder) ReplicateStartChildWorkflowExecutionFailedEvent(
-	event *types.HistoryEvent,
-) error {
-
-	attributes := event.StartChildWorkflowExecutionFailedEventAttributes
-	initiatedID := attributes.GetInitiatedEventID()
-
-	return e.DeletePendingChildExecution(initiatedID)
-}
-
-func (e *mutableStateBuilder) AddChildWorkflowExecutionCompletedEvent(
-	initiatedID int64,
-	childExecution *types.WorkflowExecution,
-	attributes *types.WorkflowExecutionCompletedEventAttributes,
-) (*types.HistoryEvent, error) {
-
-	opTag := tag.WorkflowActionChildWorkflowCompleted
-	if err := e.checkMutability(opTag); err != nil {
-		return nil, err
-	}
-
-	ci, ok := e.GetChildExecutionInfo(initiatedID)
-	if !ok || ci.StartedID == common.EmptyEventID {
-		e.logWarn(mutableStateInvalidHistoryActionMsg, opTag,
-			tag.WorkflowEventID(e.GetNextEventID()),
-			tag.ErrorTypeInvalidHistoryAction,
-			tag.Bool(ok),
-			tag.WorkflowInitiatedID(initiatedID))
-		return nil, e.createInternalServerError(opTag)
-	}
-
-	childDomainName, err := GetChildExecutionDomainName(ci, e.shard.GetDomainCache(), e.GetDomainEntry())
-	if err != nil {
-		return nil, err
-	}
-	workflowType := &types.WorkflowType{
-		Name: ci.WorkflowTypeName,
-	}
-
-	event := e.hBuilder.AddChildWorkflowExecutionCompletedEvent(
-		childDomainName,
-		childExecution,
-		workflowType,
-		ci.InitiatedID,
-		ci.StartedID,
-		attributes,
-	)
-	if err := e.ReplicateChildWorkflowExecutionCompletedEvent(event); err != nil {
-		return nil, err
-	}
-	return event, nil
-}
-
-func (e *mutableStateBuilder) ReplicateChildWorkflowExecutionCompletedEvent(
-	event *types.HistoryEvent,
-) error {
-
-	attributes := event.ChildWorkflowExecutionCompletedEventAttributes
-	initiatedID := attributes.GetInitiatedEventID()
-
-	return e.DeletePendingChildExecution(initiatedID)
-}
-
-func (e *mutableStateBuilder) AddChildWorkflowExecutionFailedEvent(
-	initiatedID int64,
-	childExecution *types.WorkflowExecution,
-	attributes *types.WorkflowExecutionFailedEventAttributes,
-) (*types.HistoryEvent, error) {
-
-	opTag := tag.WorkflowActionChildWorkflowFailed
-	if err := e.checkMutability(opTag); err != nil {
-		return nil, err
-	}
-
-	ci, ok := e.GetChildExecutionInfo(initiatedID)
-	if !ok || ci.StartedID == common.EmptyEventID {
-		e.logWarn(mutableStateInvalidHistoryActionMsg,
-			tag.WorkflowEventID(e.GetNextEventID()),
-			tag.ErrorTypeInvalidHistoryAction,
-			tag.Bool(!ok),
-			tag.WorkflowInitiatedID(initiatedID))
-		return nil, e.createInternalServerError(opTag)
-	}
-
-	domainName, err := GetChildExecutionDomainName(ci, e.shard.GetDomainCache(), e.GetDomainEntry())
-	if err != nil {
-		return nil, err
-	}
-	workflowType := &types.WorkflowType{
-		Name: ci.WorkflowTypeName,
-	}
-
-	event := e.hBuilder.AddChildWorkflowExecutionFailedEvent(
-		domainName,
-		childExecution,
-		workflowType,
-		ci.InitiatedID,
-		ci.StartedID,
-		attributes,
-	)
-	if err := e.ReplicateChildWorkflowExecutionFailedEvent(event); err != nil {
-		return nil, err
-	}
-	return event, nil
-}
-
-func (e *mutableStateBuilder) ReplicateChildWorkflowExecutionFailedEvent(
-	event *types.HistoryEvent,
-) error {
-
-	attributes := event.ChildWorkflowExecutionFailedEventAttributes
-	initiatedID := attributes.GetInitiatedEventID()
-
-	return e.DeletePendingChildExecution(initiatedID)
-}
-
-func (e *mutableStateBuilder) AddChildWorkflowExecutionCanceledEvent(
-	initiatedID int64,
-	childExecution *types.WorkflowExecution,
-	attributes *types.WorkflowExecutionCanceledEventAttributes,
-) (*types.HistoryEvent, error) {
-
-	opTag := tag.WorkflowActionChildWorkflowCanceled
-	if err := e.checkMutability(opTag); err != nil {
-		return nil, err
-	}
-
-	ci, ok := e.GetChildExecutionInfo(initiatedID)
-	if !ok || ci.StartedID == common.EmptyEventID {
-		e.logWarn(mutableStateInvalidHistoryActionMsg, opTag,
-			tag.WorkflowEventID(e.GetNextEventID()),
-			tag.ErrorTypeInvalidHistoryAction,
-			tag.Bool(ok),
-			tag.WorkflowInitiatedID(initiatedID))
-		return nil, e.createInternalServerError(opTag)
-	}
-
-	domainName, err := GetChildExecutionDomainName(ci, e.shard.GetDomainCache(), e.GetDomainEntry())
-	if err != nil {
-		return nil, err
-	}
-	workflowType := &types.WorkflowType{
-		Name: ci.WorkflowTypeName,
-	}
-
-	event := e.hBuilder.AddChildWorkflowExecutionCanceledEvent(
-		domainName,
-		childExecution,
-		workflowType,
-		ci.InitiatedID,
-		ci.StartedID,
-		attributes,
-	)
-	if err := e.ReplicateChildWorkflowExecutionCanceledEvent(event); err != nil {
-		return nil, err
-	}
-	return event, nil
-}
-
-func (e *mutableStateBuilder) ReplicateChildWorkflowExecutionCanceledEvent(
-	event *types.HistoryEvent,
-) error {
-
-	attributes := event.ChildWorkflowExecutionCanceledEventAttributes
-	initiatedID := attributes.GetInitiatedEventID()
-
-	return e.DeletePendingChildExecution(initiatedID)
-}
-
-func (e *mutableStateBuilder) AddChildWorkflowExecutionTerminatedEvent(
-	initiatedID int64,
-	childExecution *types.WorkflowExecution,
-	attributes *types.WorkflowExecutionTerminatedEventAttributes,
-) (*types.HistoryEvent, error) {
-
-	opTag := tag.WorkflowActionChildWorkflowTerminated
-	if err := e.checkMutability(opTag); err != nil {
-		return nil, err
-	}
-
-	ci, ok := e.GetChildExecutionInfo(initiatedID)
-	if !ok || ci.StartedID == common.EmptyEventID {
-		e.logWarn(mutableStateInvalidHistoryActionMsg, opTag,
-			tag.WorkflowEventID(e.GetNextEventID()),
-			tag.ErrorTypeInvalidHistoryAction,
-			tag.Bool(ok),
-			tag.WorkflowInitiatedID(initiatedID))
-		return nil, e.createInternalServerError(opTag)
-	}
-
-	domainName, err := GetChildExecutionDomainName(ci, e.shard.GetDomainCache(), e.GetDomainEntry())
-	if err != nil {
-		return nil, err
-	}
-	workflowType := &types.WorkflowType{
-		Name: ci.WorkflowTypeName,
-	}
-
-	event := e.hBuilder.AddChildWorkflowExecutionTerminatedEvent(
-		domainName,
-		childExecution,
-		workflowType,
-		ci.InitiatedID,
-		ci.StartedID,
-		attributes,
-	)
-	if err := e.ReplicateChildWorkflowExecutionTerminatedEvent(event); err != nil {
-		return nil, err
-	}
-	return event, nil
-}
-
-func (e *mutableStateBuilder) ReplicateChildWorkflowExecutionTerminatedEvent(
-	event *types.HistoryEvent,
-) error {
-
-	attributes := event.ChildWorkflowExecutionTerminatedEventAttributes
-	initiatedID := attributes.GetInitiatedEventID()
-
-	return e.DeletePendingChildExecution(initiatedID)
-}
-
-func (e *mutableStateBuilder) AddChildWorkflowExecutionTimedOutEvent(
-	initiatedID int64,
-	childExecution *types.WorkflowExecution,
-	attributes *types.WorkflowExecutionTimedOutEventAttributes,
-) (*types.HistoryEvent, error) {
-
-	opTag := tag.WorkflowActionChildWorkflowTimedOut
-	if err := e.checkMutability(opTag); err != nil {
-		return nil, err
-	}
-
-	ci, ok := e.GetChildExecutionInfo(initiatedID)
-	if !ok || ci.StartedID == common.EmptyEventID {
-		e.logWarn(mutableStateInvalidHistoryActionMsg, opTag,
-			tag.WorkflowEventID(e.GetNextEventID()),
-			tag.ErrorTypeInvalidHistoryAction,
-			tag.Bool(ok),
-			tag.WorkflowInitiatedID(initiatedID))
-		return nil, e.createInternalServerError(opTag)
-	}
-
-	domainName, err := GetChildExecutionDomainName(ci, e.shard.GetDomainCache(), e.GetDomainEntry())
-	if err != nil {
-		return nil, err
-	}
-	workflowType := &types.WorkflowType{
-		Name: ci.WorkflowTypeName,
-	}
-
-	event := e.hBuilder.AddChildWorkflowExecutionTimedOutEvent(
-		domainName,
-		childExecution,
-		workflowType,
-		ci.InitiatedID,
-		ci.StartedID,
-		attributes,
-	)
-	if err := e.ReplicateChildWorkflowExecutionTimedOutEvent(event); err != nil {
-		return nil, err
-	}
-	return event, nil
-}
-
-func (e *mutableStateBuilder) ReplicateChildWorkflowExecutionTimedOutEvent(
-	event *types.HistoryEvent,
-) error {
-
-	attributes := event.ChildWorkflowExecutionTimedOutEventAttributes
-	initiatedID := attributes.GetInitiatedEventID()
-
-	return e.DeletePendingChildExecution(initiatedID)
-}
-
-func (e *mutableStateBuilder) RetryActivity(
-	ai *persistence.ActivityInfo,
-	failureReason string,
-	failureDetails []byte,
-) (bool, error) {
-
-	opTag := tag.WorkflowActionActivityTaskRetry
-	if err := e.checkMutability(opTag); err != nil {
-		return false, err
-	}
-
-	if !ai.HasRetryPolicy || ai.CancelRequested {
-		return false, nil
-	}
-
-	now := e.timeSource.Now()
-
-	backoffInterval := getBackoffInterval(
-		now,
-		ai.ExpirationTime,
-		ai.Attempt,
-		ai.MaximumAttempts,
-		ai.InitialInterval,
-		ai.MaximumInterval,
-		ai.BackoffCoefficient,
-		failureReason,
-		ai.NonRetriableErrors,
-	)
-	if backoffInterval == backoff.NoBackoff {
-		return false, nil
-	}
-
-	// a retry is needed, update activity info for next retry
-	ai.Version = e.GetCurrentVersion()
-	ai.Attempt++
-	ai.ScheduledTime = now.Add(backoffInterval) // update to next schedule time
-	ai.StartedID = common.EmptyEventID
-	ai.RequestID = ""
-	ai.StartedTime = time.Time{}
-	ai.TimerTaskStatus = TimerTaskStatusNone
-	ai.LastFailureReason = failureReason
-	ai.LastWorkerIdentity = ai.StartedIdentity
-	ai.LastFailureDetails = failureDetails
-
-	if err := e.taskGenerator.GenerateActivityRetryTasks(
-		ai.ScheduleID,
-	); err != nil {
-		return false, err
-	}
-
-	e.updateActivityInfos[ai.ScheduleID] = ai
-	e.syncActivityTasks[ai.ScheduleID] = struct{}{}
-	return true, nil
 }
 
 // TODO mutable state should generate corresponding transfer / timer tasks according to
@@ -4902,4 +2235,18 @@ func (e *mutableStateBuilder) logDataInconsistency() {
 		tag.WorkflowID(workflowID),
 		tag.WorkflowRunID(runID),
 	)
+}
+
+func mergeMapOfByteArray(
+	current map[string][]byte,
+	upsert map[string][]byte,
+) map[string][]byte {
+
+	if current == nil {
+		current = make(map[string][]byte)
+	}
+	for k, v := range upsert {
+		current[k] = v
+	}
+	return current
 }
