+++ service/history/execution/mutable_state_builder_methods_child_workflow.go
@@ -1,0 +1,569 @@
+package execution
+
+import (
+	"context"
+	"fmt"
+	"github.com/uber/cadence/common"
+	"github.com/uber/cadence/common/backoff"
+	"github.com/uber/cadence/common/log/tag"
+	"github.com/uber/cadence/common/persistence"
+	"github.com/uber/cadence/common/types"
+	"time"
+)
+
+// GetChildExecutionInfo gives details about a child execution that is currently in progress.
+func (e *mutableStateBuilder) GetChildExecutionInfo(
+	initiatedEventID int64,
+) (*persistence.ChildExecutionInfo, bool) {
+
+	ci, ok := e.pendingChildExecutionInfoIDs[initiatedEventID]
+	return ci, ok
+}
+
+// GetChildExecutionInitiatedEvent reads out the ChildExecutionInitiatedEvent from mutable state for in-progress child
+// executions
+func (e *mutableStateBuilder) GetChildExecutionInitiatedEvent(
+	ctx context.Context,
+	initiatedEventID int64,
+) (*types.HistoryEvent, error) {
+
+	ci, ok := e.pendingChildExecutionInfoIDs[initiatedEventID]
+	if !ok {
+		return nil, ErrMissingChildWorkflowInfo
+	}
+
+	// Needed for backward compatibility reason
+	if ci.InitiatedEvent != nil {
+		return ci.InitiatedEvent, nil
+	}
+
+	currentBranchToken, err := e.GetCurrentBranchToken()
+	if err != nil {
+		return nil, err
+	}
+	initiatedEvent, err := e.eventsCache.GetEvent(
+		ctx,
+		e.shard.GetShardID(),
+		e.executionInfo.DomainID,
+		e.executionInfo.WorkflowID,
+		e.executionInfo.RunID,
+		ci.InitiatedEventBatchID,
+		ci.InitiatedID,
+		currentBranchToken,
+	)
+	if err != nil {
+		// do not return the original error
+		// since original error can be of type entity not exists
+		// which can cause task processing side to fail silently
+		// However, if the error is a persistence transient error,
+		// we return the original error, because we fail to get
+		// the event because of failure from database
+		if persistence.IsTransientError(err) {
+			return nil, err
+		}
+		return nil, ErrMissingChildWorkflowInitiatedEvent
+	}
+	return initiatedEvent, nil
+}
+
+func (e *mutableStateBuilder) GetPendingChildExecutionInfos() map[int64]*persistence.ChildExecutionInfo {
+	return e.pendingChildExecutionInfoIDs
+}
+
+// DeletePendingChildExecution deletes details about a ChildExecutionInfo.
+func (e *mutableStateBuilder) DeletePendingChildExecution(
+	initiatedEventID int64,
+) error {
+
+	if _, ok := e.pendingChildExecutionInfoIDs[initiatedEventID]; ok {
+		delete(e.pendingChildExecutionInfoIDs, initiatedEventID)
+	} else {
+		e.logError(
+			fmt.Sprintf("unable to find child workflow event ID: %v in mutable state", initiatedEventID),
+			tag.ErrorTypeInvalidMutableStateAction,
+		)
+		// log data inconsistency instead of returning an error
+		e.logDataInconsistency()
+	}
+
+	delete(e.updateChildExecutionInfos, initiatedEventID)
+	e.deleteChildExecutionInfos[initiatedEventID] = struct{}{}
+	return nil
+}
+
+func (e *mutableStateBuilder) AddStartChildWorkflowExecutionInitiatedEvent(
+	decisionCompletedEventID int64,
+	createRequestID string,
+	attributes *types.StartChildWorkflowExecutionDecisionAttributes,
+) (*types.HistoryEvent, *persistence.ChildExecutionInfo, error) {
+
+	opTag := tag.WorkflowActionChildWorkflowInitiated
+	if err := e.checkMutability(opTag); err != nil {
+		return nil, nil, err
+	}
+
+	event := e.hBuilder.AddStartChildWorkflowExecutionInitiatedEvent(
+		decisionCompletedEventID,
+		attributes,
+		e.GetDomainEntry().GetInfo().Name,
+	)
+	// Write the event to cache only on active cluster
+	e.eventsCache.PutEvent(e.executionInfo.DomainID, e.executionInfo.WorkflowID, e.executionInfo.RunID,
+		event.ID, event)
+
+	ci, err := e.ReplicateStartChildWorkflowExecutionInitiatedEvent(decisionCompletedEventID, event, createRequestID)
+	if err != nil {
+		return nil, nil, err
+	}
+	return event, ci, nil
+}
+
+func (e *mutableStateBuilder) ReplicateStartChildWorkflowExecutionInitiatedEvent(
+	firstEventID int64,
+	event *types.HistoryEvent,
+	createRequestID string,
+) (*persistence.ChildExecutionInfo, error) {
+
+	initiatedEventID := event.ID
+	attributes := event.StartChildWorkflowExecutionInitiatedEventAttributes
+
+	domainID := e.GetExecutionInfo().DomainID
+	if domainName := attributes.GetDomain(); domainName != "" {
+		// domainName may still be empty if two cadence clusters are running different versions
+		var err error
+		domainID, err = e.shard.GetDomainCache().GetDomainID(domainName)
+		if err != nil {
+			return nil, err
+		}
+	}
+	ci := &persistence.ChildExecutionInfo{
+		Version:               event.Version,
+		InitiatedID:           initiatedEventID,
+		InitiatedEventBatchID: firstEventID,
+		StartedID:             common.EmptyEventID,
+		StartedWorkflowID:     attributes.GetWorkflowID(),
+		CreateRequestID:       createRequestID,
+		DomainID:              domainID,
+		// DomainName field is being deprecated
+		// DomainName:            attributes.GetDomain(),
+		WorkflowTypeName:  attributes.GetWorkflowType().GetName(),
+		ParentClosePolicy: attributes.GetParentClosePolicy(),
+	}
+
+	e.pendingChildExecutionInfoIDs[ci.InitiatedID] = ci
+	e.updateChildExecutionInfos[ci.InitiatedID] = ci
+
+	return ci, e.taskGenerator.GenerateChildWorkflowTasks(event)
+}
+
+func (e *mutableStateBuilder) AddChildWorkflowExecutionStartedEvent(
+	domain string,
+	execution *types.WorkflowExecution,
+	workflowType *types.WorkflowType,
+	initiatedID int64,
+	header *types.Header,
+) (*types.HistoryEvent, error) {
+
+	opTag := tag.WorkflowActionChildWorkflowStarted
+	if err := e.checkMutability(opTag); err != nil {
+		return nil, err
+	}
+
+	ci, ok := e.GetChildExecutionInfo(initiatedID)
+	if !ok || ci.StartedID != common.EmptyEventID {
+		e.logWarn(mutableStateInvalidHistoryActionMsg, opTag,
+			tag.WorkflowEventID(e.GetNextEventID()),
+			tag.ErrorTypeInvalidHistoryAction,
+			tag.Bool(ok),
+			tag.WorkflowInitiatedID(initiatedID))
+		return nil, e.createInternalServerError(opTag)
+	}
+
+	event := e.hBuilder.AddChildWorkflowExecutionStartedEvent(domain, execution, workflowType, initiatedID, header)
+	if err := e.ReplicateChildWorkflowExecutionStartedEvent(event); err != nil {
+		return nil, err
+	}
+	return event, nil
+}
+
+func (e *mutableStateBuilder) ReplicateChildWorkflowExecutionStartedEvent(
+	event *types.HistoryEvent,
+) error {
+
+	attributes := event.ChildWorkflowExecutionStartedEventAttributes
+	initiatedID := attributes.GetInitiatedEventID()
+
+	ci, ok := e.GetChildExecutionInfo(initiatedID)
+	if !ok {
+		e.logError(
+			"Unable to find child workflow",
+			tag.ErrorTypeInvalidMutableStateAction,
+			tag.WorkflowEventID(e.GetNextEventID()),
+			tag.WorkflowInitiatedID(initiatedID),
+		)
+		return ErrMissingChildWorkflowInfo
+	}
+	ci.StartedID = event.ID
+	ci.StartedRunID = attributes.GetWorkflowExecution().GetRunID()
+	e.updateChildExecutionInfos[ci.InitiatedID] = ci
+
+	return nil
+}
+
+func (e *mutableStateBuilder) AddStartChildWorkflowExecutionFailedEvent(
+	initiatedID int64,
+	cause types.ChildWorkflowExecutionFailedCause,
+	initiatedEventAttributes *types.StartChildWorkflowExecutionInitiatedEventAttributes,
+) (*types.HistoryEvent, error) {
+
+	opTag := tag.WorkflowActionChildWorkflowInitiationFailed
+	if err := e.checkMutability(opTag); err != nil {
+		return nil, err
+	}
+
+	ci, ok := e.GetChildExecutionInfo(initiatedID)
+	if !ok || ci.StartedID != common.EmptyEventID {
+		e.logWarn(mutableStateInvalidHistoryActionMsg, opTag,
+			tag.WorkflowEventID(e.GetNextEventID()),
+			tag.ErrorTypeInvalidHistoryAction,
+			tag.Bool(ok),
+			tag.WorkflowInitiatedID(initiatedID))
+		return nil, e.createInternalServerError(opTag)
+	}
+
+	event := e.hBuilder.AddStartChildWorkflowExecutionFailedEvent(initiatedID, cause, initiatedEventAttributes)
+	if err := e.ReplicateStartChildWorkflowExecutionFailedEvent(event); err != nil {
+		return nil, err
+	}
+	return event, nil
+}
+
+func (e *mutableStateBuilder) ReplicateStartChildWorkflowExecutionFailedEvent(
+	event *types.HistoryEvent,
+) error {
+
+	attributes := event.StartChildWorkflowExecutionFailedEventAttributes
+	initiatedID := attributes.GetInitiatedEventID()
+
+	return e.DeletePendingChildExecution(initiatedID)
+}
+
+func (e *mutableStateBuilder) AddChildWorkflowExecutionCompletedEvent(
+	initiatedID int64,
+	childExecution *types.WorkflowExecution,
+	attributes *types.WorkflowExecutionCompletedEventAttributes,
+) (*types.HistoryEvent, error) {
+
+	opTag := tag.WorkflowActionChildWorkflowCompleted
+	if err := e.checkMutability(opTag); err != nil {
+		return nil, err
+	}
+
+	ci, ok := e.GetChildExecutionInfo(initiatedID)
+	if !ok || ci.StartedID == common.EmptyEventID {
+		e.logWarn(mutableStateInvalidHistoryActionMsg, opTag,
+			tag.WorkflowEventID(e.GetNextEventID()),
+			tag.ErrorTypeInvalidHistoryAction,
+			tag.Bool(ok),
+			tag.WorkflowInitiatedID(initiatedID))
+		return nil, e.createInternalServerError(opTag)
+	}
+
+	childDomainName, err := GetChildExecutionDomainName(ci, e.shard.GetDomainCache(), e.GetDomainEntry())
+	if err != nil {
+		return nil, err
+	}
+	workflowType := &types.WorkflowType{
+		Name: ci.WorkflowTypeName,
+	}
+
+	event := e.hBuilder.AddChildWorkflowExecutionCompletedEvent(
+		childDomainName,
+		childExecution,
+		workflowType,
+		ci.InitiatedID,
+		ci.StartedID,
+		attributes,
+	)
+	if err := e.ReplicateChildWorkflowExecutionCompletedEvent(event); err != nil {
+		return nil, err
+	}
+	return event, nil
+}
+
+func (e *mutableStateBuilder) ReplicateChildWorkflowExecutionCompletedEvent(
+	event *types.HistoryEvent,
+) error {
+
+	attributes := event.ChildWorkflowExecutionCompletedEventAttributes
+	initiatedID := attributes.GetInitiatedEventID()
+
+	return e.DeletePendingChildExecution(initiatedID)
+}
+
+func (e *mutableStateBuilder) AddChildWorkflowExecutionFailedEvent(
+	initiatedID int64,
+	childExecution *types.WorkflowExecution,
+	attributes *types.WorkflowExecutionFailedEventAttributes,
+) (*types.HistoryEvent, error) {
+
+	opTag := tag.WorkflowActionChildWorkflowFailed
+	if err := e.checkMutability(opTag); err != nil {
+		return nil, err
+	}
+
+	ci, ok := e.GetChildExecutionInfo(initiatedID)
+	if !ok || ci.StartedID == common.EmptyEventID {
+		e.logWarn(mutableStateInvalidHistoryActionMsg,
+			tag.WorkflowEventID(e.GetNextEventID()),
+			tag.ErrorTypeInvalidHistoryAction,
+			tag.Bool(!ok),
+			tag.WorkflowInitiatedID(initiatedID))
+		return nil, e.createInternalServerError(opTag)
+	}
+
+	domainName, err := GetChildExecutionDomainName(ci, e.shard.GetDomainCache(), e.GetDomainEntry())
+	if err != nil {
+		return nil, err
+	}
+	workflowType := &types.WorkflowType{
+		Name: ci.WorkflowTypeName,
+	}
+
+	event := e.hBuilder.AddChildWorkflowExecutionFailedEvent(
+		domainName,
+		childExecution,
+		workflowType,
+		ci.InitiatedID,
+		ci.StartedID,
+		attributes,
+	)
+	if err := e.ReplicateChildWorkflowExecutionFailedEvent(event); err != nil {
+		return nil, err
+	}
+	return event, nil
+}
+
+func (e *mutableStateBuilder) ReplicateChildWorkflowExecutionFailedEvent(
+	event *types.HistoryEvent,
+) error {
+
+	attributes := event.ChildWorkflowExecutionFailedEventAttributes
+	initiatedID := attributes.GetInitiatedEventID()
+
+	return e.DeletePendingChildExecution(initiatedID)
+}
+
+func (e *mutableStateBuilder) AddChildWorkflowExecutionCanceledEvent(
+	initiatedID int64,
+	childExecution *types.WorkflowExecution,
+	attributes *types.WorkflowExecutionCanceledEventAttributes,
+) (*types.HistoryEvent, error) {
+
+	opTag := tag.WorkflowActionChildWorkflowCanceled
+	if err := e.checkMutability(opTag); err != nil {
+		return nil, err
+	}
+
+	ci, ok := e.GetChildExecutionInfo(initiatedID)
+	if !ok || ci.StartedID == common.EmptyEventID {
+		e.logWarn(mutableStateInvalidHistoryActionMsg, opTag,
+			tag.WorkflowEventID(e.GetNextEventID()),
+			tag.ErrorTypeInvalidHistoryAction,
+			tag.Bool(ok),
+			tag.WorkflowInitiatedID(initiatedID))
+		return nil, e.createInternalServerError(opTag)
+	}
+
+	domainName, err := GetChildExecutionDomainName(ci, e.shard.GetDomainCache(), e.GetDomainEntry())
+	if err != nil {
+		return nil, err
+	}
+	workflowType := &types.WorkflowType{
+		Name: ci.WorkflowTypeName,
+	}
+
+	event := e.hBuilder.AddChildWorkflowExecutionCanceledEvent(
+		domainName,
+		childExecution,
+		workflowType,
+		ci.InitiatedID,
+		ci.StartedID,
+		attributes,
+	)
+	if err := e.ReplicateChildWorkflowExecutionCanceledEvent(event); err != nil {
+		return nil, err
+	}
+	return event, nil
+}
+
+func (e *mutableStateBuilder) ReplicateChildWorkflowExecutionCanceledEvent(
+	event *types.HistoryEvent,
+) error {
+
+	attributes := event.ChildWorkflowExecutionCanceledEventAttributes
+	initiatedID := attributes.GetInitiatedEventID()
+
+	return e.DeletePendingChildExecution(initiatedID)
+}
+
+func (e *mutableStateBuilder) AddChildWorkflowExecutionTerminatedEvent(
+	initiatedID int64,
+	childExecution *types.WorkflowExecution,
+	attributes *types.WorkflowExecutionTerminatedEventAttributes,
+) (*types.HistoryEvent, error) {
+
+	opTag := tag.WorkflowActionChildWorkflowTerminated
+	if err := e.checkMutability(opTag); err != nil {
+		return nil, err
+	}
+
+	ci, ok := e.GetChildExecutionInfo(initiatedID)
+	if !ok || ci.StartedID == common.EmptyEventID {
+		e.logWarn(mutableStateInvalidHistoryActionMsg, opTag,
+			tag.WorkflowEventID(e.GetNextEventID()),
+			tag.ErrorTypeInvalidHistoryAction,
+			tag.Bool(ok),
+			tag.WorkflowInitiatedID(initiatedID))
+		return nil, e.createInternalServerError(opTag)
+	}
+
+	domainName, err := GetChildExecutionDomainName(ci, e.shard.GetDomainCache(), e.GetDomainEntry())
+	if err != nil {
+		return nil, err
+	}
+	workflowType := &types.WorkflowType{
+		Name: ci.WorkflowTypeName,
+	}
+
+	event := e.hBuilder.AddChildWorkflowExecutionTerminatedEvent(
+		domainName,
+		childExecution,
+		workflowType,
+		ci.InitiatedID,
+		ci.StartedID,
+		attributes,
+	)
+	if err := e.ReplicateChildWorkflowExecutionTerminatedEvent(event); err != nil {
+		return nil, err
+	}
+	return event, nil
+}
+
+func (e *mutableStateBuilder) ReplicateChildWorkflowExecutionTerminatedEvent(
+	event *types.HistoryEvent,
+) error {
+
+	attributes := event.ChildWorkflowExecutionTerminatedEventAttributes
+	initiatedID := attributes.GetInitiatedEventID()
+
+	return e.DeletePendingChildExecution(initiatedID)
+}
+
+func (e *mutableStateBuilder) AddChildWorkflowExecutionTimedOutEvent(
+	initiatedID int64,
+	childExecution *types.WorkflowExecution,
+	attributes *types.WorkflowExecutionTimedOutEventAttributes,
+) (*types.HistoryEvent, error) {
+
+	opTag := tag.WorkflowActionChildWorkflowTimedOut
+	if err := e.checkMutability(opTag); err != nil {
+		return nil, err
+	}
+
+	ci, ok := e.GetChildExecutionInfo(initiatedID)
+	if !ok || ci.StartedID == common.EmptyEventID {
+		e.logWarn(mutableStateInvalidHistoryActionMsg, opTag,
+			tag.WorkflowEventID(e.GetNextEventID()),
+			tag.ErrorTypeInvalidHistoryAction,
+			tag.Bool(ok),
+			tag.WorkflowInitiatedID(initiatedID))
+		return nil, e.createInternalServerError(opTag)
+	}
+
+	domainName, err := GetChildExecutionDomainName(ci, e.shard.GetDomainCache(), e.GetDomainEntry())
+	if err != nil {
+		return nil, err
+	}
+	workflowType := &types.WorkflowType{
+		Name: ci.WorkflowTypeName,
+	}
+
+	event := e.hBuilder.AddChildWorkflowExecutionTimedOutEvent(
+		domainName,
+		childExecution,
+		workflowType,
+		ci.InitiatedID,
+		ci.StartedID,
+		attributes,
+	)
+	if err := e.ReplicateChildWorkflowExecutionTimedOutEvent(event); err != nil {
+		return nil, err
+	}
+	return event, nil
+}
+
+func (e *mutableStateBuilder) ReplicateChildWorkflowExecutionTimedOutEvent(
+	event *types.HistoryEvent,
+) error {
+
+	attributes := event.ChildWorkflowExecutionTimedOutEventAttributes
+	initiatedID := attributes.GetInitiatedEventID()
+
+	return e.DeletePendingChildExecution(initiatedID)
+}
+
+func (e *mutableStateBuilder) RetryActivity(
+	ai *persistence.ActivityInfo,
+	failureReason string,
+	failureDetails []byte,
+) (bool, error) {
+
+	opTag := tag.WorkflowActionActivityTaskRetry
+	if err := e.checkMutability(opTag); err != nil {
+		return false, err
+	}
+
+	if !ai.HasRetryPolicy || ai.CancelRequested {
+		return false, nil
+	}
+
+	now := e.timeSource.Now()
+
+	backoffInterval := getBackoffInterval(
+		now,
+		ai.ExpirationTime,
+		ai.Attempt,
+		ai.MaximumAttempts,
+		ai.InitialInterval,
+		ai.MaximumInterval,
+		ai.BackoffCoefficient,
+		failureReason,
+		ai.NonRetriableErrors,
+	)
+	if backoffInterval == backoff.NoBackoff {
+		return false, nil
+	}
+
+	// a retry is needed, update activity info for next retry
+	ai.Version = e.GetCurrentVersion()
+	ai.Attempt++
+	ai.ScheduledTime = now.Add(backoffInterval) // update to next schedule time
+	ai.StartedID = common.EmptyEventID
+	ai.RequestID = ""
+	ai.StartedTime = time.Time{}
+	ai.TimerTaskStatus = TimerTaskStatusNone
+	ai.LastFailureReason = failureReason
+	ai.LastWorkerIdentity = ai.StartedIdentity
+	ai.LastFailureDetails = failureDetails
+
+	if err := e.taskGenerator.GenerateActivityRetryTasks(
+		ai.ScheduleID,
+	); err != nil {
+		return false, err
+	}
+
+	e.updateActivityInfos[ai.ScheduleID] = ai
+	e.syncActivityTasks[ai.ScheduleID] = struct{}{}
+	return true, nil
+}
