+++ service/history/execution/mutable_state_builder_methods_signal.go
@@ -1,0 +1,232 @@
+package execution
+
+import (
+	"fmt"
+	"github.com/uber/cadence/common/log/tag"
+	"github.com/uber/cadence/common/persistence"
+	"github.com/uber/cadence/common/types"
+)
+
+func (e *mutableStateBuilder) IsSignalRequested(
+	requestID string,
+) bool {
+
+	if _, ok := e.pendingSignalRequestedIDs[requestID]; ok {
+		return true
+	}
+	return false
+}
+
+// GetSignalInfo get details about a signal request that is currently in progress.
+func (e *mutableStateBuilder) GetSignalInfo(
+	initiatedEventID int64,
+) (*persistence.SignalInfo, bool) {
+
+	ri, ok := e.pendingSignalInfoIDs[initiatedEventID]
+	return ri, ok
+}
+
+func (e *mutableStateBuilder) GetPendingSignalExternalInfos() map[int64]*persistence.SignalInfo {
+	return e.pendingSignalInfoIDs
+
+}
+
+func (e *mutableStateBuilder) AddWorkflowExecutionSignaled(
+	signalName string,
+	input []byte,
+	identity string,
+	requestID string,
+) (*types.HistoryEvent, error) {
+
+	opTag := tag.WorkflowActionWorkflowSignaled
+	if err := e.checkMutability(opTag); err != nil {
+		return nil, err
+	}
+
+	event := e.hBuilder.AddWorkflowExecutionSignaledEvent(signalName, input, identity, requestID)
+	if err := e.ReplicateWorkflowExecutionSignaled(event); err != nil {
+		return nil, err
+	}
+	return event, nil
+}
+
+func (e *mutableStateBuilder) ReplicateWorkflowExecutionSignaled(
+	event *types.HistoryEvent,
+) error {
+
+	// Increment signal count in mutable state for this workflow execution
+	e.executionInfo.SignalCount++
+	e.insertWorkflowRequest(persistence.WorkflowRequest{
+		RequestID:   event.WorkflowExecutionSignaledEventAttributes.RequestID,
+		Version:     event.Version,
+		RequestType: persistence.WorkflowRequestTypeSignal,
+	})
+	return nil
+}
+
+func (e *mutableStateBuilder) AddExternalWorkflowExecutionSignaled(
+	initiatedID int64,
+	domain string,
+	workflowID string,
+	runID string,
+	control []byte,
+) (*types.HistoryEvent, error) {
+
+	opTag := tag.WorkflowActionExternalWorkflowSignalRequested
+	if err := e.checkMutability(opTag); err != nil {
+		return nil, err
+	}
+
+	_, ok := e.GetSignalInfo(initiatedID)
+	if !ok {
+		e.logWarn(mutableStateInvalidHistoryActionMsg, opTag,
+			tag.WorkflowEventID(e.GetNextEventID()),
+			tag.ErrorTypeInvalidHistoryAction,
+			tag.WorkflowInitiatedID(initiatedID))
+		return nil, e.createInternalServerError(opTag)
+	}
+
+	event := e.hBuilder.AddExternalWorkflowExecutionSignaled(initiatedID, domain, workflowID, runID, control)
+	if err := e.ReplicateExternalWorkflowExecutionSignaled(event); err != nil {
+		return nil, err
+	}
+	return event, nil
+}
+
+func (e *mutableStateBuilder) ReplicateExternalWorkflowExecutionSignaled(
+	event *types.HistoryEvent,
+) error {
+
+	initiatedID := event.ExternalWorkflowExecutionSignaledEventAttributes.GetInitiatedEventID()
+
+	return e.DeletePendingSignal(initiatedID)
+}
+
+func (e *mutableStateBuilder) AddSignalExternalWorkflowExecutionFailedEvent(
+	decisionTaskCompletedEventID int64,
+	initiatedID int64,
+	domain string,
+	workflowID string,
+	runID string,
+	control []byte,
+	cause types.SignalExternalWorkflowExecutionFailedCause,
+) (*types.HistoryEvent, error) {
+
+	opTag := tag.WorkflowActionExternalWorkflowSignalFailed
+	if err := e.checkMutability(opTag); err != nil {
+		return nil, err
+	}
+
+	_, ok := e.GetSignalInfo(initiatedID)
+	if !ok {
+		e.logWarn(mutableStateInvalidHistoryActionMsg, opTag,
+			tag.WorkflowEventID(e.GetNextEventID()),
+			tag.ErrorTypeInvalidHistoryAction,
+			tag.WorkflowInitiatedID(initiatedID))
+		return nil, e.createInternalServerError(opTag)
+	}
+
+	event := e.hBuilder.AddSignalExternalWorkflowExecutionFailedEvent(decisionTaskCompletedEventID, initiatedID, domain,
+		workflowID, runID, control, cause)
+	if err := e.ReplicateSignalExternalWorkflowExecutionFailedEvent(event); err != nil {
+		return nil, err
+	}
+	return event, nil
+}
+
+func (e *mutableStateBuilder) ReplicateSignalExternalWorkflowExecutionFailedEvent(
+	event *types.HistoryEvent,
+) error {
+
+	initiatedID := event.SignalExternalWorkflowExecutionFailedEventAttributes.GetInitiatedEventID()
+
+	return e.DeletePendingSignal(initiatedID)
+}
+
+func (e *mutableStateBuilder) AddSignalRequested(
+	requestID string,
+) {
+
+	if e.pendingSignalRequestedIDs == nil {
+		e.pendingSignalRequestedIDs = make(map[string]struct{})
+	}
+	if e.updateSignalRequestedIDs == nil {
+		e.updateSignalRequestedIDs = make(map[string]struct{})
+	}
+	e.pendingSignalRequestedIDs[requestID] = struct{}{} // add requestID to set
+	e.updateSignalRequestedIDs[requestID] = struct{}{}
+}
+
+func (e *mutableStateBuilder) DeleteSignalRequested(
+	requestID string,
+) {
+
+	delete(e.pendingSignalRequestedIDs, requestID)
+	delete(e.updateSignalRequestedIDs, requestID)
+	e.deleteSignalRequestedIDs[requestID] = struct{}{}
+}
+
+func (e *mutableStateBuilder) AddSignalExternalWorkflowExecutionInitiatedEvent(
+	decisionCompletedEventID int64,
+	signalRequestID string,
+	request *types.SignalExternalWorkflowExecutionDecisionAttributes,
+) (*types.HistoryEvent, *persistence.SignalInfo, error) {
+
+	opTag := tag.WorkflowActionExternalWorkflowSignalInitiated
+	if err := e.checkMutability(opTag); err != nil {
+		return nil, nil, err
+	}
+
+	event := e.hBuilder.AddSignalExternalWorkflowExecutionInitiatedEvent(decisionCompletedEventID, request)
+	si, err := e.ReplicateSignalExternalWorkflowExecutionInitiatedEvent(decisionCompletedEventID, event, signalRequestID)
+	if err != nil {
+		return nil, nil, err
+	}
+	return event, si, nil
+}
+
+func (e *mutableStateBuilder) ReplicateSignalExternalWorkflowExecutionInitiatedEvent(
+	firstEventID int64,
+	event *types.HistoryEvent,
+	signalRequestID string,
+) (*persistence.SignalInfo, error) {
+
+	// TODO: Consider also writing signalRequestID to history event
+	initiatedEventID := event.ID
+	attributes := event.SignalExternalWorkflowExecutionInitiatedEventAttributes
+	si := &persistence.SignalInfo{
+		Version:               event.Version,
+		InitiatedEventBatchID: firstEventID,
+		InitiatedID:           initiatedEventID,
+		SignalRequestID:       signalRequestID,
+		SignalName:            attributes.GetSignalName(),
+		Input:                 attributes.Input,
+		Control:               attributes.Control,
+	}
+
+	e.pendingSignalInfoIDs[si.InitiatedID] = si
+	e.updateSignalInfos[si.InitiatedID] = si
+
+	return si, e.taskGenerator.GenerateSignalExternalTasks(event)
+}
+
+// DeletePendingSignal deletes details about a SignalInfo
+func (e *mutableStateBuilder) DeletePendingSignal(
+	initiatedEventID int64,
+) error {
+
+	if _, ok := e.pendingSignalInfoIDs[initiatedEventID]; ok {
+		delete(e.pendingSignalInfoIDs, initiatedEventID)
+	} else {
+		e.logError(
+			fmt.Sprintf("unable to find signal external workflow event ID: %v in mutable state", initiatedEventID),
+			tag.ErrorTypeInvalidMutableStateAction,
+		)
+		// log data inconsistency instead of returning an error
+		e.logDataInconsistency()
+	}
+
+	delete(e.updateSignalInfos, initiatedEventID)
+	e.deleteSignalInfos[initiatedEventID] = struct{}{}
+	return nil
+}
