+++ service/history/execution/mutable_state_builder_methods_timedout.go
@@ -1,0 +1,41 @@
+package execution
+
+import (
+	"github.com/uber/cadence/common/log/tag"
+	"github.com/uber/cadence/common/persistence"
+	"github.com/uber/cadence/common/types"
+)
+
+func (e *mutableStateBuilder) AddTimeoutWorkflowEvent(
+	firstEventID int64,
+) (*types.HistoryEvent, error) {
+
+	opTag := tag.WorkflowActionWorkflowTimeout
+	if err := e.checkMutability(opTag); err != nil {
+		return nil, err
+	}
+
+	event := e.hBuilder.AddTimeoutWorkflowEvent()
+	if err := e.ReplicateWorkflowExecutionTimedoutEvent(firstEventID, event); err != nil {
+		return nil, err
+	}
+	return event, nil
+}
+
+func (e *mutableStateBuilder) ReplicateWorkflowExecutionTimedoutEvent(
+	firstEventID int64,
+	event *types.HistoryEvent,
+) error {
+
+	if err := e.UpdateWorkflowStateCloseStatus(
+		persistence.WorkflowStateCompleted,
+		persistence.WorkflowCloseStatusTimedOut,
+	); err != nil {
+		return err
+	}
+	e.executionInfo.CompletionEventBatchID = firstEventID // Used when completion event needs to be loaded from database
+	e.ClearStickyness()
+	e.writeEventToCache(event)
+
+	return e.taskGenerator.GenerateWorkflowCloseTasks(event, e.config.WorkflowDeletionJitterRange(e.domainEntry.GetInfo().Name))
+}
