+++ service/history/execution/mutable_state_builder_methods_timer.go
@@ -1,0 +1,261 @@
+package execution
+
+import (
+	"fmt"
+	"github.com/uber/cadence/common/log/tag"
+	"github.com/uber/cadence/common/persistence"
+	"github.com/uber/cadence/common/types"
+	"time"
+)
+
+func (e *mutableStateBuilder) GetPendingTimerInfos() map[string]*persistence.TimerInfo {
+	return e.pendingTimerInfoIDs
+}
+
+func (e *mutableStateBuilder) AddTimerStartedEvent(
+	decisionCompletedEventID int64,
+	request *types.StartTimerDecisionAttributes,
+) (*types.HistoryEvent, *persistence.TimerInfo, error) {
+
+	opTag := tag.WorkflowActionTimerStarted
+	if err := e.checkMutability(opTag); err != nil {
+		return nil, nil, err
+	}
+
+	timerID := request.GetTimerID()
+	_, ok := e.GetUserTimerInfo(timerID)
+	if ok {
+		e.logWarn(mutableStateInvalidHistoryActionMsg, opTag,
+			tag.WorkflowEventID(e.GetNextEventID()),
+			tag.ErrorTypeInvalidHistoryAction,
+			tag.WorkflowTimerID(timerID))
+		return nil, nil, e.createCallerError(opTag)
+	}
+
+	event := e.hBuilder.AddTimerStartedEvent(decisionCompletedEventID, request)
+	ti, err := e.ReplicateTimerStartedEvent(event)
+	if err != nil {
+		return nil, nil, err
+	}
+	return event, ti, err
+}
+
+func (e *mutableStateBuilder) ReplicateTimerStartedEvent(
+	event *types.HistoryEvent,
+) (*persistence.TimerInfo, error) {
+
+	attributes := event.TimerStartedEventAttributes
+	timerID := attributes.GetTimerID()
+
+	startToFireTimeout := attributes.GetStartToFireTimeoutSeconds()
+	fireTimeout := time.Duration(startToFireTimeout) * time.Second
+	// TODO: Time skew need to be taken in to account.
+	expiryTime := time.Unix(0, event.GetTimestamp()).Add(fireTimeout) // should use the event time, not now
+	ti := &persistence.TimerInfo{
+		Version:    event.Version,
+		TimerID:    timerID,
+		ExpiryTime: expiryTime,
+		StartedID:  event.ID,
+		TaskStatus: TimerTaskStatusNone,
+	}
+
+	e.pendingTimerInfoIDs[ti.TimerID] = ti
+	e.pendingTimerEventIDToID[ti.StartedID] = ti.TimerID
+	e.updateTimerInfos[ti.TimerID] = ti
+
+	return ti, nil
+}
+
+func (e *mutableStateBuilder) AddTimerFiredEvent(
+	timerID string,
+) (*types.HistoryEvent, error) {
+
+	opTag := tag.WorkflowActionTimerFired
+	if err := e.checkMutability(opTag); err != nil {
+		return nil, err
+	}
+
+	timerInfo, ok := e.GetUserTimerInfo(timerID)
+	if !ok {
+		e.logWarn(mutableStateInvalidHistoryActionMsg, opTag,
+			tag.WorkflowEventID(e.GetNextEventID()),
+			tag.ErrorTypeInvalidHistoryAction,
+			tag.WorkflowTimerID(timerID))
+		return nil, e.createInternalServerError(opTag)
+	}
+
+	// Timer is running.
+	event := e.hBuilder.AddTimerFiredEvent(timerInfo.StartedID, timerID)
+	if err := e.ReplicateTimerFiredEvent(event); err != nil {
+		return nil, err
+	}
+	return event, nil
+}
+
+func (e *mutableStateBuilder) ReplicateTimerFiredEvent(
+	event *types.HistoryEvent,
+) error {
+
+	attributes := event.TimerFiredEventAttributes
+	timerID := attributes.GetTimerID()
+
+	return e.DeleteUserTimer(timerID)
+}
+
+func (e *mutableStateBuilder) AddTimerCanceledEvent(
+	decisionCompletedEventID int64,
+	attributes *types.CancelTimerDecisionAttributes,
+	identity string,
+) (*types.HistoryEvent, error) {
+
+	opTag := tag.WorkflowActionTimerCanceled
+	if err := e.checkMutability(opTag); err != nil {
+		return nil, err
+	}
+
+	var timerStartedID int64
+	timerID := attributes.GetTimerID()
+	ti, ok := e.GetUserTimerInfo(timerID)
+	if !ok {
+		// if timer is not running then check if it has fired in the mutable state.
+		// If so clear the timer from the mutable state. We need to check both the
+		// bufferedEvents and the history builder
+		timerFiredEvent := e.checkAndClearTimerFiredEvent(timerID)
+		if timerFiredEvent == nil {
+			e.logWarn(mutableStateInvalidHistoryActionMsg, opTag,
+				tag.WorkflowEventID(e.GetNextEventID()),
+				tag.ErrorTypeInvalidHistoryAction,
+				tag.WorkflowTimerID(timerID))
+			return nil, e.createCallerError(opTag)
+		}
+		timerStartedID = timerFiredEvent.TimerFiredEventAttributes.GetStartedEventID()
+	} else {
+		timerStartedID = ti.StartedID
+	}
+
+	// Timer is running.
+	event := e.hBuilder.AddTimerCanceledEvent(timerStartedID, decisionCompletedEventID, timerID, identity)
+	if ok {
+		if err := e.ReplicateTimerCanceledEvent(event); err != nil {
+			return nil, err
+		}
+	}
+	return event, nil
+}
+
+func (e *mutableStateBuilder) ReplicateTimerCanceledEvent(
+	event *types.HistoryEvent,
+) error {
+
+	attributes := event.TimerCanceledEventAttributes
+	timerID := attributes.GetTimerID()
+
+	return e.DeleteUserTimer(timerID)
+}
+
+func (e *mutableStateBuilder) AddCancelTimerFailedEvent(
+	decisionCompletedEventID int64,
+	attributes *types.CancelTimerDecisionAttributes,
+	identity string,
+) (*types.HistoryEvent, error) {
+
+	opTag := tag.WorkflowActionTimerCancelFailed
+	if err := e.checkMutability(opTag); err != nil {
+		return nil, err
+	}
+
+	// No Operation: We couldn't cancel it probably TIMER_ID_UNKNOWN
+	timerID := attributes.GetTimerID()
+	return e.hBuilder.AddCancelTimerFailedEvent(timerID, decisionCompletedEventID,
+		timerCancellationMsgTimerIDUnknown, identity), nil
+}
+
+// GetUserTimerInfo gives details about a user timer.
+func (e *mutableStateBuilder) GetUserTimerInfo(
+	timerID string,
+) (*persistence.TimerInfo, bool) {
+
+	timerInfo, ok := e.pendingTimerInfoIDs[timerID]
+	return timerInfo, ok
+}
+
+// UpdateUserTimer updates the user timer in progress.
+func (e *mutableStateBuilder) UpdateUserTimer(
+	ti *persistence.TimerInfo,
+) error {
+
+	timerID, ok := e.pendingTimerEventIDToID[ti.StartedID]
+	if !ok {
+		e.logError(
+			fmt.Sprintf("unable to find timer event ID: %v in mutable state", ti.StartedID),
+			tag.ErrorTypeInvalidMutableStateAction,
+		)
+		return ErrMissingTimerInfo
+	}
+
+	if _, ok := e.pendingTimerInfoIDs[timerID]; !ok {
+		e.logError(
+			fmt.Sprintf("unable to find timer ID: %v in mutable state", timerID),
+			tag.ErrorTypeInvalidMutableStateAction,
+		)
+		return ErrMissingTimerInfo
+	}
+
+	e.pendingTimerInfoIDs[ti.TimerID] = ti
+	e.updateTimerInfos[ti.TimerID] = ti
+	return nil
+}
+
+// DeleteUserTimer deletes an user timer.
+func (e *mutableStateBuilder) DeleteUserTimer(
+	timerID string,
+) error {
+
+	if timerInfo, ok := e.pendingTimerInfoIDs[timerID]; ok {
+		delete(e.pendingTimerInfoIDs, timerID)
+
+		if _, ok = e.pendingTimerEventIDToID[timerInfo.StartedID]; ok {
+			delete(e.pendingTimerEventIDToID, timerInfo.StartedID)
+		} else {
+			e.logError(
+				fmt.Sprintf("unable to find timer event ID: %v in mutable state", timerID),
+				tag.ErrorTypeInvalidMutableStateAction,
+			)
+			// log data inconsistency instead of returning an error
+			e.logDataInconsistency()
+		}
+	} else {
+		e.logError(
+			fmt.Sprintf("unable to find timer ID: %v in mutable state", timerID),
+			tag.ErrorTypeInvalidMutableStateAction,
+		)
+		// log data inconsistency instead of returning an error
+		e.logDataInconsistency()
+	}
+
+	delete(e.updateTimerInfos, timerID)
+	e.deleteTimerInfos[timerID] = struct{}{}
+	return nil
+}
+
+func checkAndClearTimerFiredEvent(
+	events []*types.HistoryEvent,
+	timerID string,
+) ([]*types.HistoryEvent, *types.HistoryEvent) {
+	// go over all history events. if we find a timer fired event for the given
+	// timerID, clear it
+	timerFiredIdx := -1
+	for idx, event := range events {
+		if event.GetEventType() == types.EventTypeTimerFired &&
+			event.GetTimerFiredEventAttributes().GetTimerID() == timerID {
+			timerFiredIdx = idx
+			break
+		}
+	}
+	if timerFiredIdx == -1 {
+		return events, nil
+	}
+
+	timerEvent := events[timerFiredIdx]
+	return append(events[:timerFiredIdx], events[timerFiredIdx+1:]...), timerEvent
+}
